### 2017-09-13 #152 @boshek
New submission: tidyhydat
-   What does this package do? (explain in 50 words or less): `tidyhydat` provides functions to extract historical and real-time national hydrometric data from Water Survey of Canada data sources

```
Package: tidyhydat
Title: Extract and Tidy Canadian Hydrometric Data
Version: 0.2.9
Authors@R: c(person("Sam", "Albers", email = "sam.albers@gov.bc.ca", role = c("aut", "cre"),
           ), person("David", "Hutchinson", email = "david.hutchinson@canada.ca", role = "ctb"),
           person("Province of British Columbia", role = "cph"))
Description: tidyhydat provides functions to extract historical and real-time national hydrometric
  data from Water Survey of Canada data sources (http://dd.weather.gc.ca/hydrometric/csv/ and
  http://collaboration.cmc.ec.gc.ca/cmc/hydrometrics/www/) and then apply tidy data principles.
License: Apache License (== 2.0) | file LICENSE
    readr (>= 1.1.1),
    lubridate (>= 1.6.0),
    tidyr (>= 0.7.1),
    DBI (>= 0.7),
    RSQLite (>= 2.0),
    tibble (>= 1.3.4),
    httr (>= 1.3.1)
    tidyverse,
    dbplyr,
```

-   URL for the package (the development repository, not a stylized html page): https://github.com/bcgov/tidyhydat

- Please indicate which category or categories from our [package fit policies](https://github.com/ropensci/onboarding/blob/master/policies.md#package-fit) this package falls under ***and why**:

    - data retrieval: `tidyhydat` retrieves data from the Water Survey of Canada *datamart* via the `download_realtime_dd()` function from http://dd.weather.gc.ca/hydrometric/csv/. `download_realtime_ws` provides access via the `httr` package to a new password protected web service provided by Environment and Climate Change Canada (ECCC). Differences between the download realtime functions are outlined [here](https://github.com/bcgov/tidyhydat/blob/master/vignettes/tidyhydat.Rmd#compare-download_realtime_ws-and-download_realtime_dd)
    - data extraction: The majority of `tidyhydat`'s exported functions provide easy access to ECCC's publicly available and open-licenced HYDAT sqlite3 database (http://collaboration.cmc.ec.gc.ca/cmc/hydrometrics/www/) of historical river and lake levels in Canada. HYDAT is updated quarterly and is distributed via the [Canadian Open Government Licence](https://github.com/bcgov/tidyhydat/blob/master/data-raw/HYDAT_internal_data/LICENSE.OGL-CAN-2.0). HYDAT is not included with this package because it is regularly updated and because the file size is prohibitively large (>1GB) for inclusion in an R package. Rather `tidyhydat` provides the `download_hydat()` function that downloads it for the user.

    - Anyone who is interested in making use of the Water Survey of Canada data sources in R including industry, government and academics.

    - A search of the term "hydat" on github reveals two R packages that provide similar (but not all the same) functionality:
    - [HYDAT](https://github.com/CentreForHydrology/HYDAT): the original author of **HYDAT** has contributed to this package and further development or maintenance of **HYDAT** is not imminent.
    - [hydatr](https://github.com/paleolimbot/hydatr) was developed at approximately the same time as `tidyhydat`.
    - `tidyhydat` documentation is more comprehensive than both packages.
    - Through the realtime functions `tidyhydat` provides added functionality not present in both packages. Neither `HYDAT` nor `hydatr` provide any functionality to access the web service and `hydatr` does not access any real-time datamart data.
    - An express goal of `tidyhydat` is to provide data in a tidy format. This conceptual data science goal provides a clear objective that is missing from both other packages.

Taken from the list reviewers as folks that might be appropriate:
   - Joseph Stachelek: @jsta
   - Jeffrey Hollister: @jhollist
   - Alison Appling: @aappling-usgs
   - Laura DeCicco: @ldecicco-USGS
