### 2017-09-14 dbe7817
Update documentation and examples
