+++ R/SED_DLY_LOADS.R
@@ -24,11 +24,7 @@
 #'
 #' @examples
 #' \donttest{
-#' SED_DLY_LOADS(STATION_NUMBER = c("08MH024","08MH001"), hydat_path = "H:/Hydat.sqlite3",
-#' start_date = "1996-01-01", end_date = "2000-01-01")
-#'
 #' SED_DLY_LOADS(PROV_TERR_STATE_LOC = "PE", hydat_path = "H:/Hydat.sqlite3")
-#'
 #'           }
 #'
 #' @family HYDAT functions
