+++ R/SED_MONTHLY_LOADS.R
@@ -25,8 +25,6 @@
 #'
 #' @examples
 #' \donttest{
-#' SED_MONTHLY_LOADS(STATION_NUMBER = c("02JE013","08MF005"), hydat_path = "H:/Hydat.sqlite3",
-#' start_date = "1996-01-01", end_date = "2000-01-01")
 #'
 #' SED_MONTHLY_LOADS(PROV_TERR_STATE_LOC = "PE", hydat_path = "H:/Hydat.sqlite3")
 #'           }
