+++ man/SED_DLY_LOADS.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/SED_DLY_LOAD.R
+% Please edit documentation in R/SED_DLY_LOADS.R
 \name{SED_DLY_LOADS}
 \alias{SED_DLY_LOADS}
 \title{Extract daily sediment load information from the HYDAT database}
@@ -35,11 +35,7 @@ That is a large vector for \code{SED_DLY_LOADS}.
 }
 \examples{
 \donttest{
-SED_DLY_LOADS(STATION_NUMBER = c("08MH024","08MH001"), hydat_path = "H:/Hydat.sqlite3",
-start_date = "1996-01-01", end_date = "2000-01-01")
-
 SED_DLY_LOADS(PROV_TERR_STATE_LOC = "PE", hydat_path = "H:/Hydat.sqlite3")
-
           }
 
 }
