+++ man/SED_MONTHLY_SUSCON.Rd
@@ -36,9 +36,6 @@ Provides wrapper to turn the SED_MONTHLY_SUSCON table in HYDAT into a tidy data 
 }
 \examples{
 \donttest{
-SED_MONTHLY_SUSCON(STATION_NUMBER = c("02JE013","08MF005"), hydat_path = "H:/Hydat.sqlite3",
-start_date = "1996-01-01", end_date = "2000-01-01")
-
 SED_MONTHLY_SUSCON(PROV_TERR_STATE_LOC = "PE", hydat_path = "H:/Hydat.sqlite3")
           }
 }
