### 2017-09-14 05f3946
move from 1:nrow to seq_along
