### 2017-09-14 #152 @sckott
### Editor checks:

---

#### Editor comments

again, thanks for your submission @boshek

Thanks for including `goodpractice` output

Seeking reviewers now.

---

Reviewers:
Due date:
