### 2017-09-25 c4504ee
confirm UTC time setting
