### 2017-09-25 70ec10d
confirm UTC time setting. Closes #32
