+++ man/download_realtime_ws.Rd
@@ -21,7 +21,7 @@ download_realtime_ws(STATION_NUMBER, parameters = c(46, 16, 52, 47, 8, 5, 41,
 \item{token}{generate by \code{get_ws_token()}}
 }
 \value{
-Time is return as UTC for consistency
+Time is return as UTC for consistency.
 }
 \description{
 Download realtime data from the ECCC web service
