### 2017-09-25 e172fa5
Added test for realtime_network_meta and moved to httr to download.
