### 2017-09-25 e815d8a
Improved error handling for download_realtime_dd. Closes #35 & #36
