### 2017-09-25 31784ab
download_realtime_dd uses httr now for retrieval. Small speed gain.
