### 2017-09-25 7b67001
removed need for almost all @import statement by referencing them all directly. Closes #34
