+++ tests/testthat/test_realtime_network_meta.R
@@ -4,3 +4,4 @@ context("Testing realtime_network_meta ")
    expect_silent(realtime_network_meta(PROV_TERR_STATE_LOC = "BC"))
  })
 
+ 
