+++ R/ANNUAL_STATISTICS.R
@@ -73,17 +73,17 @@ ANNUAL_STATISTICS <- function(hydat_path=NULL, STATION_NUMBER =NULL, PROV_TERR_S
 
   ## TODO: Figure out how to do this in fewer steps
   ## Mean tibble
-  as_mean <- select(annual_statistics, STATION_NUMBER, DATA_TYPE, YEAR, MEAN)
-  as_mean <- gather(as_mean, SUM_STAT, Value, -STATION_NUMBER, -DATA_TYPE, -YEAR)
+  as_mean <- dplyr::select(annual_statistics, STATION_NUMBER, DATA_TYPE, YEAR, MEAN)
+  as_mean <- tidyr::gather(as_mean, SUM_STAT, Value, -STATION_NUMBER, -DATA_TYPE, -YEAR)
 
   ## Min tibble
-  as_min <- select(annual_statistics, STATION_NUMBER, DATA_TYPE, YEAR, MIN_MONTH, MIN_DAY, MIN, MIN_SYMBOL)
-  as_min <- gather(as_min, SUM_STAT, Value, -STATION_NUMBER, -DATA_TYPE, -YEAR, -MIN_MONTH, -MIN_DAY, -MIN_SYMBOL)
+  as_min <- dplyr::select(annual_statistics, STATION_NUMBER, DATA_TYPE, YEAR, MIN_MONTH, MIN_DAY, MIN, MIN_SYMBOL)
+  as_min <- tidyr::gather(as_min, SUM_STAT, Value, -STATION_NUMBER, -DATA_TYPE, -YEAR, -MIN_MONTH, -MIN_DAY, -MIN_SYMBOL)
   colnames(as_min) <- gsub("MIN_", "", names(as_min))
 
   ## Max tibble
-  as_max <- select(annual_statistics, STATION_NUMBER, DATA_TYPE, YEAR, MAX_MONTH, MAX_DAY, MAX, MAX_SYMBOL)
-  as_max <- gather(as_max, SUM_STAT, Value, -STATION_NUMBER, -DATA_TYPE, -YEAR, -MAX_MONTH, -MAX_DAY, -MAX_SYMBOL)
+  as_max <- dplyr::select(annual_statistics, STATION_NUMBER, DATA_TYPE, YEAR, MAX_MONTH, MAX_DAY, MAX, MAX_SYMBOL)
+  as_max <- tidyr::gather(as_max, SUM_STAT, Value, -STATION_NUMBER, -DATA_TYPE, -YEAR, -MAX_MONTH, -MAX_DAY, -MAX_SYMBOL)
   colnames(as_max) <- gsub("MAX_", "", names(as_max))
 
   ## bind into 1 dataframe and by year and join in the symbol
@@ -100,7 +100,7 @@ ANNUAL_STATISTICS <- function(hydat_path=NULL, STATION_NUMBER =NULL, PROV_TERR_S
   annual_statistics <- dplyr::left_join(annual_statistics, DATA_TYPES, by = c("DATA_TYPE"))
 
   ## Clean up the variables
-  annual_statistics <- select(annual_statistics, STATION_NUMBER, DATA_TYPE_EN, YEAR:Value, Date, SYMBOL_EN)
+  annual_statistics <- dplyr::select(annual_statistics, STATION_NUMBER, DATA_TYPE_EN, YEAR:Value, Date, SYMBOL_EN)
 
   ## Rename to tidyhydat format
   colnames(annual_statistics) <- c("STATION_NUMBER", "Parameter", "Year", "Sum_stat", "Value", "Date", "Symbol")
