+++ R/utils.R
@@ -71,7 +71,7 @@ AGENCY_LIST <- function(hydat_path=NULL) {
   on.exit(DBI::dbDisconnect(hydat_con))
 
   agency_list <- dplyr::tbl(hydat_con, "AGENCY_LIST") %>%
-    collect()
+    dplyr::collect()
 
   agency_list
 }
@@ -110,7 +110,7 @@ REGIONAL_OFFICE_LIST <- function(hydat_path=NULL) {
 
 
   regional_office_list <- dplyr::tbl(hydat_con, "REGIONAL_OFFICE_LIST") %>%
-    collect()
+    dplyr::collect()
   
   regional_office_list
 }
@@ -146,7 +146,7 @@ DATUM_LIST <- function(hydat_path=NULL) {
   on.exit(DBI::dbDisconnect(hydat_con))
 
   datum_list <- dplyr::tbl(hydat_con, "DATUM_LIST") %>%
-    collect()
+    dplyr::collect()
 
   datum_list
 }

[…]
