### 2017-09-25 303f126
Fixed error when calling with ::. Closes #31
