### 2017-09-25 fcd44f2
reformatted code to minimize some spacing
