+++ R/ANNUAL_INSTANT_PEAKS.R
@@ -103,6 +103,6 @@ ANNUAL_INSTANT_PEAKS <- function(hydat_path = NULL, STATION_NUMBER = NULL,
   } else {
     message("All station successfully retrieved")
   }
-
+  
   aip
 }
