### 2017-09-25 9d1da94
To reduce overhead,  has been added as an argument. Closes #33
