+++ R/DLY_FLOWS.R
@@ -19,6 +19,8 @@
 #' @inheritParams STATIONS
 #' @param start_date Leave blank if all dates are required. Date format needs to be in YYYY-MM-DD. Date is inclusive.
 #' @param end_date Leave blank if all dates are required. Date format needs to be in YYYY-MM-DD. Date is inclusive.
+#' @param symbol_output Set whether the raw code, or the \code{english} or the \code{french} translations are outputted. Default 
+#'   value is \code{code}. 
 #'
 #' @return A tibble of daily levels
 #'
@@ -37,11 +39,13 @@
 
 
 
-DLY_FLOWS <- function(hydat_path=NULL, STATION_NUMBER = NULL, PROV_TERR_STATE_LOC = NULL, start_date ="ALL", end_date = "ALL") {
+DLY_FLOWS <- function(hydat_path=NULL, STATION_NUMBER = NULL, 
+                      PROV_TERR_STATE_LOC = NULL, start_date ="ALL", end_date = "ALL",
+                      symbol_output = "code") {
   if (!is.null(STATION_NUMBER) && STATION_NUMBER == "ALL") {
     stop("Deprecated behaviour.Omit the STATION_NUMBER = \"ALL\" argument. See ?DLY_FLOWS for examples.")
   }
-
+  
   if (start_date == "ALL" & end_date == "ALL") {
     message("No start and end dates specified. All dates available will be returned.")
   } else {
@@ -50,24 +54,24 @@ DLY_FLOWS <- function(hydat_path=NULL, STATION_NUMBER = NULL, PROV_TERR_STATE_LO
     start_year <- lubridate::year(start_date)
     start_month <- lubridate::month(start_date)
     start_day <- lubridate::day(start_date)
-
+    
     ## End
     end_year <- lubridate::year(end_date)
     end_month <- lubridate::month(end_date)
     end_day <- lubridate::day(end_date)
   }
-
+  
   ## Check date is in the right format
   if (start_date != "ALL" | end_date != "ALL") {
     if (is.na(as.Date(start_date, format = "%Y-%m-%d")) | is.na(as.Date(end_date, format = "%Y-%m-%d"))) {
       stop("Invalid date format. Dates need to be in YYYY-MM-DD format")
     }
-
+    
     if (start_date > end_date) {
       stop("start_date is after end_date. Try swapping values.")
     }
   }
-
+  
   if (is.null(hydat_path)) {
     hydat_path <- Sys.getenv("hydat")
     if (is.na(hydat_path)) {
@@ -75,27 +79,27 @@ DLY_FLOWS <- function(hydat_path=NULL, STATION_NUMBER = NULL, PROV_TERR_STATE_LO
            in your .Renviron file. See ?tidyhydat for more documentation.")
     }
   }
-
-
+  
+  
   ## Read in database
   hydat_con <- DBI::dbConnect(RSQLite::SQLite(), hydat_path)
   
   on.exit(DBI::dbDisconnect(hydat_con))
-
+  
   ## Determine which stations we are querying
   stns <- station_choice(hydat_con, STATION_NUMBER, PROV_TERR_STATE_LOC)
-
-
+  
+  
   ## Data manipulations to make it "tidy"
   dly_flows <- dplyr::tbl(hydat_con, "DLY_FLOWS")
   dly_flows <- dplyr::filter(dly_flows, STATION_NUMBER %in% stns)
-
+  
   ## Do the initial subset to take advantage of dbplyr only issuing sql query when it has too
[…]
