### 2017-09-25 c168d4d
Updated NEWS and DESCRIPTION to v0.2.9
