### 2017-09-25 #152 @boshek
NOTE: Version update since original submission. `goodpractice` output is unchanged and the DESCRIPTION file above have updated accordingly. Changes are catalogued in the [NEWS.md](https://github.com/bcgov/tidyhydat/blob/master/NEWS.md) file in the tidyhydat repo.
