### 2017-09-27 4f84b3e
add tidyhydat hex sticker
