### 2017-10-02 b72acc9
update paper
