### 2017-10-11 2ca1ad2
fixed get_ws_token example
