+++ paper/paper.bib
@@ -20,6 +20,17 @@
   publisher = {Foundation for Open Access Statistics},
 }
 
+@article{wilson2014best,
+  title={Best practices for scientific computing},
+  author={Wilson, Greg and Aruliah, Dhavide A and Brown, C Titus and Hong, Neil P Chue and Davis, Matt and Guy, Richard T and Haddock, Steven HD and Huff, Kathryn D and Mitchell, Ian M and Plumbley, Mark D and others},
+  journal={PLoS biology},
+  volume={12},
+  number={1},
+  pages={e1001745},
+  year={2014},
+  publisher={Public Library of Science}
+}
+
 @Manual{RCore,
   title = {R: A Language and Environment for Statistical Computing},
   author = {{R Core Team}},
@@ -36,5 +47,12 @@
   note = {R package version 1.1.1},
   url = {https://CRAN.R-project.org/package=tidyverse},
 }
+
+  @Manual{alberstidyhydat,
+    title = {tidyhydat: Extract and Tidy Canadian Hydrometric Data},
+    author = {Sam Albers},
+    note = {R package version 0.2.9},
+    url = {https://github.com/bcgov/tidyhydat},
+  }
 
 @Comment{jabref-meta: databaseType:bibtex;}
