### 2017-10-12 44877c4
update documentation for download_hydat to reflect arg name
