### 2017-10-16 #152 @lawinslow
>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

#### Overall

The `tidyhydat` package offers interfaces to Environment and Climate Change Canada stream and river data.
It primarily interfaces with a downloadable database, "HYDAT", that contains a great deal of historical data.
It also provides functions for accessing some real-time data through select web interfaces.

Overall, the scientific computing world requires more tools like these. Many a student, reseracher, or manager
has had to re-create the wheel to access available data and providing such interfaces goes a long way to
accelerating and streamlining such data use.

I do have a few reservations and suggestions about the current implementation of `tidyhydat` that I think, if addressed,
would improve the package functionality and enable broader use.

One issue I encountered was the database download, with personal handling of the database path, is awkward. It seems that much of the work
of downloading the database, and then storing the path could be automated. The storage location could be supported by
the `rappdirs` package. There are existing examples of such file downloading and use, such as [getlandsat](https://github.com/ropenscilabs/getlandsat),
[hydrolinks](https://github.com/lawinslow/hydrolinks), and [LAGOS](https://github.com/cont-limno/LAGOS) to name a few
with various implementations. This would make it easier for users to quickly get started using the data and eliminate
the need to muck around in the Renviron file.

Another overarching issue (which is somewhat subjective) is to the outside reviewer, the choice of function capitalization
seems arbitrary and not helpful. While the choices of function style are subject to personal preferences, I would suggest
the usage of consistent capitalization scheme, preferrably a lowercase-based one. I am personally especially partial to
the consistent prefix-style function naming, which is used in a number of [rOpenSci](https://github.com/ropenscilabs/getlandsat/)
and other [related](https://github.com/paleolimbot/hydatr) packages. It eases function discovery as well as makes packages
seem more unified and consistent.

First thing I noticed were the typos and grammatical errors on the Github README. I'd suggest reviewing it for
grammatical issues if it will be the github face of tidyhydat (noted mistakes in first two bullet points).

I would suggest you build and commit your Rd files to github. Otherwise, devtools::install_github
will end up installing a package with no documentation. Took me 5 minutes of trying to get documentation before
I realized why no functions were documented.

Line 29 zzz.R I would drop the #' comments which results in R trying to build an Rd file for .onLoad.

Lastly. I would consider revising your tidyhydat-package documentation. This is frequently the first interaction
a user might have with the package. Currently it is very focused on the data format and output style (tidy),
not the data itself. This documentation (along with a Vignette) would be an excellent place to better give an overview
of the data available. What data are availble through this package? Stream/river level, flow, and it seems, sediment
flux estimates at daily and monthly timescales (also real-time in some cases)?

On the overview side, I'd recommend, if possible, pointing to further HYDAT documentation. A lot of the documentation here
relies on the user's knowledge of the HYDAT database. For example, much of the function documentation refers to the functions
as "Provides wrapper to turn the XXX table in HYDAT into a tidy data frame." which implies knowledge of the HYDAT database.

There seems to be no indication of the units for any of the data. The only unit I could find is in the DATA_TYPES
table, though that was only for Sediment. This is something I feel is important for data packages such as this as units
are critical to understanding the data, and are often the one piece of information I find myself spending excessive time
hunting for (or trying to reverse engineer).

Lastly, a natural question that comes to my mind is, "are there other data"? How about temperature? Or Nutrients? Or
spatial data that reflect the local watersheds and flowlines.

One thing about the examples. Using devtools I can run them, but a bunch of them are hard-coded
to use h:/Hydat.sqlite3. It would be better if these examples followed the style outlined in the github
README which uses an .Renviron variable to set the path (which would then allow anyone to quickly use/run examples).

Example for `SED_MONTHLY_LOADS` is not working. I played around a bit with it but couldn't get it to run. This
may require better output of empty return values as it seems like it should be ok that a query returns no data.
```
> SED_MONTHLY_LOADS(PROV_TERR_STATE_LOC = "PE", hydat_path = "H:/Hydat.sqlite3", start_date = "2001-01-01", end_date = "2011-01-01")

 Error: !length(data) || is_named(data) is not TRUE
11.stop(msg, call. = FALSE, domain = NA)
10.stopifnot(!length(data) || is_named(data))
9.env_bind_impl(env, dots_list(...))
8.child_env(env_, ...)
7.env_bury(overscope_top, !(!(!data)))
6.vars_select_eval(.vars, quos)
5.tidyselect::vars_select(names(data), !(!(!quos)))
4.unname(tidyselect::vars_select(names(data), !(!(!quos))))
3.gather.data.frame(sed_monthly_loads, variable, temp, -(STATION_NUMBER:NO_DAYS))
2.tidyr::gather(sed_monthly_loads, variable, temp, -(STATION_NUMBER:NO_DAYS)) at SED_MONTHLY_LOADS.R#107
1.SED_MONTHLY_LOADS(PROV_TERR_STATE_LOC = "PE", hydat_path = "H:/Hydat.sqlite3",
    start_date = "2001-01-01", end_date = "2011-01-01")
```

I found that this same function failed on the testthat unit tests as well.

Overall, I look forward to the tidyhydat package being available on CRAN and using its functionality to
integrate hydrologic data at large scales.
