+++ README.Rmd
@@ -15,13 +15,26 @@ knitr::opts_chunk$set(
 ```
 
 # tidyhydat <img src="img/tidyhydat.png" align="right" />
+
+
   
-<div id="devex-badge"><a rel="Delivery" href="https://github.com/BCDevExchange/docs/blob/master/discussion/projectstates.md"><img alt="In production, but maybe in Alpha or Beta. Intended to persist and be supported." style="border-width:0" src="https://assets.bcdevexchange.org/images/badges/delivery.svg" title="In production, but maybe in Alpha or Beta. Intended to persist and be supported." /></a> [![Travis-CI Build Status](https://travis-ci.org/bcgov/tidyhydat.svg?branch=master)](https://travis-ci.org/bcgov/tidyhydat) [![](https://badges.ropensci.org/152_status.svg)](https://github.com/ropensci/onboarding/issues/152) </div>
+<div id="devex-badge"><a rel="Exploration" href="https://github.com/BCDevExchange/docs/blob/master/discussion/projectstates.md"><img alt="Being designed and built, but in the lab. May change, disappear, or be buggy." style="border-width:0" src="https://assets.bcdevexchange.org/images/badges/exploration.svg" title="Being designed and built, but in the lab. May change, disappear, or be buggy." /></a> [![Travis-CI Build Status](https://travis-ci.org/bcgov/tidyhydat.svg?branch=master)](https://travis-ci.org/bcgov/tidyhydat) [![](https://badges.ropensci.org/152_status.svg)](https://github.com/ropensci/onboarding/issues/152) </div>
+
+## Project Status
+
+This package is under active development. The package is currently undergoing a review by [rOpenSci](https://ropensci.org/). You can see the progress of this review here: https://github.com/ropensci/onboarding/issues/152. As a result breaking changes to `tidyhydat` functions will be made. To install the most recent version prior to these changes you can use this command:
+
+```{r, eval = FALSE}
+devtools::install_github("bcgov/tidyhydat@v0.2.9")
+```
+However, you should strongly consider migrating to the updated `tidyhydat` once the reveiw process is complete. 
+
+## What does it do?
 
 Here is a list of what `tidyhydat` does:
   
-- Provide function that access each tables in the HYDAT database and return tidy data.
-- Keep functions are low-level as possible. For example, for daily flows, the `DLY_FLOWS()` function queries the database, *tidies* the data and returns the data. 
+- Provide function that access each table in the HYDAT database and return tidy data.
+- Keep functions as low-level as possible. For example, for daily flows, the `DLY_FLOWS()` function queries the database, *tidies* the data and returns the data. 
 - Provide functions that access Environment and Climate Change Canada's real-time hydrometric data source. 
 - Provide functions that search full station lists and aid in generating station vectors
   
@@ -36,7 +49,7 @@ remotes::install_github("bcgov/tidyhydat")
   
 Then to load the package you need to using the `library()` function. When you install `tidyhydat`, several other packages will be installed as well. One of those packages, `dplyr`, is useful for data manipulations and is used regularly here. Even though `dplyr` is installed alongside `tidyhydat`, it is helpful to load it by itself as there are many useful functions contained within `dplyr`. A helpful `dplyr` tutorial can be found [here](https://cran.r-project.org/web/packages/dplyr/vignettes/dplyr.html).
   
-```{r, eval = TRUE, echo=TRUE, message=FALSE}
+```{r, eval = TRUE, echo=TRUE, message=FALSE, warnig=FALSE}
 library(tidyhydat)
 library(dplyr)
 ```
@@ -102,9 +115,7 @@ search_number("08MF")
 ```
   
   
-## Project Status
 
-This package is under active development.
 
 ## Getting Help or Reporting an Issue
 
