### 2017-10-20 82ad8f2
fixed small bug in SED_MONTHLY_LOADS; addresses but doesn't totally close #51
