### 2017-10-20 8432c1c
added Sys.sleep(1); aleviated get_ws_token failures; closes #43
