### 2017-10-20 22cdb86
new NEWS format
