### 2017-10-20 676ee96
fixed small bug causing check() flag
