### 2017-10-20 14ec458
removed #' from .onload. Closes #47
