### 2017-10-20 2565aa7
update NEWS
