+++ NEWS.md
@@ -10,7 +10,8 @@ tidyhydat 0.3.0 ()
 
 ### BUG FIXES
 
-  * Fixed failing behaviour of get_ws_token (bcgov/tidyhydat#43)
+  * Fixed failing behaviour of `get_ws_token` (bcgov/tidyhydat#43)
+  * No longer trying to build .Rd file for `.onload` (bcgov/tidyhydat#47)
 
 ### DEPRECATED AND DEFUNCT
 
