### 2017-10-20 17c73b8
fixed httr:content parsing error. Closes #42
