### 2017-10-20 699c82e
update NEWS
