+++ NEWS.md
@@ -12,6 +12,7 @@ tidyhydat 0.3.0 ()
 
   * Fixed failing behaviour of `get_ws_token` (bcgov/tidyhydat#43)
   * No longer trying to build .Rd file for `.onload` (bcgov/tidyhydat#47)
+  * Fixed httr:content parsing error in `download_realtime_ws`. Closes #42"
 
 ### DEPRECATED AND DEFUNCT
 
