### 2017-10-20 1adfc61
add informative error message for solo missing station instance
