+++ R/DLY_FLOWS.R
@@ -102,6 +102,10 @@ DLY_FLOWS <- function(hydat_path=NULL, STATION_NUMBER = NULL,
   
   dly_flows <- dplyr::select(dly_flows, STATION_NUMBER, YEAR, MONTH, NO_DAYS, dplyr::contains("FLOW"))
   dly_flows <- dplyr::collect(dly_flows)
+  
+  if(is.data.frame(dly_flows) && nrow(dly_flows)==0)
+    {stop("This station is not present in HYDAT")}
+  
   dly_flows <- tidyr::gather(dly_flows, variable, temp, -(STATION_NUMBER:NO_DAYS))
   dly_flows <- dplyr::mutate(dly_flows, DAY = as.numeric(gsub("FLOW|FLOW_SYMBOL", "", variable)))
   dly_flows <- dplyr::mutate(dly_flows, variable = gsub("[0-9]+", "", variable))
