### 2017-10-20 98c079e
FULL_MONTH evaluating to logical. Closes #51
