### 2017-10-20 53c9610
adding rappdirs to imports and using to generate download path for download_hydat(). Closes #44
