+++ DESCRIPTION
@@ -20,7 +20,8 @@ Imports:
     DBI (>= 0.7),
     RSQLite (>= 2.0),
     tibble (>= 1.3.4),
-    httr (>= 1.3.1)
+    httr (>= 1.3.1),
+    rappdirs (>= 0.3.1)
 Suggests:
     tidyverse,
     dbplyr,
