### 2017-10-23 eecb56b
wrapped download_realtime_ws examples in dontrun
