### 2017-10-23 2755de4
wrapped download_realtime_ws examples in dontrun. Closes #52
