### 2017-10-23 997f00c
Update NEWS to close #52
