### 2017-10-23 #152 @boshek
@lawinslow @ldecicco-USGS @sckott

Two questions:
1. Give that I have incorporated Luke's suggestion to `rappdirs` and am looking to standardize documentation from a user perspective is it better to provide examples that user the tiny database like this which the user can use right away:
    -  `DLY_FLOWS(STATION_NUMBER = c("08LA001","08NL071"),
                      hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat"))`
Or would it better to have examples that user the full hydat database like this:
   -  `DLY_FLOWS(STATION_NUMBER = c("08LA001","08NL071"))`
Knowing that users would have to download the full database before being able to use those functions.

   Any opinions on which approach is less annoying? The "get started right away" approach or the "learn it how I am going to use it" approach?

2. The recommendation was to switch from `DLY_FLOWS()` to `hydat_dly_flows()`. However I did not consider the length of other function. In particular this function `ANNUAL_INSTANT_PEAKS()` becomes `hydat_annual_instant_peaks()` (similar situation with `SED_MONTHLY_SUSCON()`). This result in, at least in my estimation, a function name that is too long. Are there any objections to omitting the `hydat_*` portion of the function name to save space?
