### 2017-10-23 #152 @sckott
hi @boshek , i can't decipher this text above

> is it better to provide examples that user the tiny database like this which the user can use right away

can you clarify?
