### 2017-10-23 #152 @boshek
@sckott Updated the offending lines. Does that help?
