### 2017-10-23 3234c22
added skip on cran test; tidied up some error messages and removed time limit for queries. Closes #53
