+++ tests/testthat/test_download_realtime.R
@@ -21,6 +21,7 @@ context("Testing download_realtime functions")
  
  
  test_that("download_realtime_dd returns the correct data header", {
+   skip_on_cran()
    expect_identical(
      colnames(download_realtime_dd(STATION_NUMBER = "08MF005", PROV_TERR_STATE_LOC = "BC")),
      c("STATION_NUMBER", "PROV_TERR_STATE_LOC", "Date", "Parameter", "Value", "Grade", "Symbol", "Code")
@@ -28,14 +29,17 @@ context("Testing download_realtime functions")
 })
 
 test_that("download_realtime_dd can download stations from multiple provinces using PROV_TERR_STATE_LOC", {
+  skip_on_cran()
   download_realtime_dd(PROV_TERR_STATE_LOC = c("QC", "PE"))
 })
 
 
 test_that("download_realtime_dd can download stations from multiple provinces using STATION_NUMBER", {
+  skip_on_cran()
   expect_error(download_realtime_dd(STATION_NUMBER = c("01CD005", "08MF005")), regexp = NA)
 })
 
 test_that("When STATION_NUMBER is ALL there is an error", {
+  skip_on_cran()
   expect_error(download_realtime_dd(STATION_NUMBER = "ALL"))
 })
