+++ NEWS.md
@@ -9,6 +9,8 @@ tidyhydat 0.3.0 ()
   * Adding `rappdirs` to imports and using to generate download path for `download_hydat()` (bcgov/tidyhydat#44)
   * Add informative error message for a single missing station input (bcgov/tidyhydat#38)
   * `FULL MONTH` evaluate to a logic (bcgov/tidyhydat#51)
+  * All download tests are skipped on cran (bcgov/tidyhydat#53)
+  * Removed time limit for `download_realtime_ws()` with some documentation on actual limits. (https://github.com/bcgov/tidyhydat/commit/3234c2246c97fed5860e8dfb9adc3d6f0aa503fe)
 
 ### BUG FIXES
 
@@ -19,6 +21,7 @@ tidyhydat 0.3.0 ()
   * Wrapped download_realtime_ws examples in dontrun (bcgov/tidyhydat#52)
 
 ### DEPRECATED AND DEFUNCT
+  * 
 
 
 
