### 2017-10-24 229a7fb
added rappdirs for path settings; first step for tidytools #54; better examples #50
