+++ R/SED_MONTHLY_SUSCON.R
@@ -25,7 +25,7 @@
 #'
 #' @examples
 #' \donttest{
-#' SED_MONTHLY_SUSCON(PROV_TERR_STATE_LOC = "PE", hydat_path = "H:/Hydat.sqlite3")
+#' SED_MONTHLY_SUSCON(STATION_NUMBER = "08MF005")
 #'           }
 #' @family HYDAT functions
 #' @source HYDAT
@@ -33,9 +33,16 @@
 
 
 
-SED_MONTHLY_SUSCON <- function(hydat_path=NULL, STATION_NUMBER = NULL, PROV_TERR_STATE_LOC = NULL, start_date ="ALL", end_date = "ALL") {
+SED_MONTHLY_SUSCON <- function(STATION_NUMBER = NULL, 
+                               hydat_path = paste0(rappdirs::user_data_dir(),"\\Hydat.sqlite3"), 
+                               PROV_TERR_STATE_LOC = NULL, start_date ="ALL", end_date = "ALL") {
   if (!is.null(STATION_NUMBER) && STATION_NUMBER == "ALL") {
     stop("Deprecated behaviour.Omit the STATION_NUMBER = \"ALL\" argument. See ?SED_MONTHLY_SUSCON for examples.")
+  }
+  
+  ## Check if hydat is present
+  if (!file.exists(hydat_path)){
+    stop(paste0("No Hydat.sqlite3 found at ",rappdirs::user_data_dir(),". Run download_hydat() to download the database."))
   }
 
   if (start_date == "ALL" & end_date == "ALL") {
@@ -61,14 +68,6 @@ SED_MONTHLY_SUSCON <- function(hydat_path=NULL, STATION_NUMBER = NULL, PROV_TERR
 
     if (start_date > end_date) {
       stop("start_date is after end_date. Try swapping values.")
-    }
-  }
-
-  if (is.null(hydat_path)) {
-    hydat_path <- Sys.getenv("hydat")
-    if (is.na(hydat_path)) {
-      stop("No Hydat.sqlite3 path set either in this function or 
-           in your .Renviron file. See ?tidyhydat for more documentation.")
     }
   }
 
