### 2017-10-24 73cce1b
replace function names in tests
