+++ tests/testthat/test_DLY_LEVELS.R
@@ -1,36 +1,36 @@
-context("Testing DLY_LEVELS")
+context("Testing hy_daily_levels")
 
-test_that("DLY_LEVELS accepts single and multiple province arguments", {
+test_that("hy_daily_levels accepts single and multiple province arguments", {
   stns <- "08MF005"
-  expect_identical(unique(DLY_LEVELS(STATION_NUMBER = stns, 
+  expect_identical(unique(hy_daily_levels(STATION_NUMBER = stns, 
                                      hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat"))$STATION_NUMBER), stns)
-  expect_identical(length(unique(DLY_LEVELS(STATION_NUMBER = c("08MF005", "05AA008"), hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat"))$STATION_NUMBER)), length(c("08NM083", "08NE102")))
+  expect_identical(length(unique(hy_daily_levels(STATION_NUMBER = c("08MF005", "05AA008"), hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat"))$STATION_NUMBER)), length(c("08NM083", "08NE102")))
 })
 
 
-test_that("DLY_LEVELS accepts single and multiple province arguments", {
-  expect_true(nrow(DLY_LEVELS(PROV_TERR_STATE_LOC = "BC", hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat"))) >= 1)
-  expect_true(nrow(DLY_LEVELS(PROV_TERR_STATE_LOC = c("YT", "BC"), hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat"))) >= 1)
+test_that("hy_daily_levels accepts single and multiple province arguments", {
+  expect_true(nrow(hy_daily_levels(PROV_TERR_STATE_LOC = "BC", hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat"))) >= 1)
+  expect_true(nrow(hy_daily_levels(PROV_TERR_STATE_LOC = c("YT", "BC"), hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat"))) >= 1)
 })
 
-test_that("DLY_LEVELS produces an error when a province is not specified correctly", {
-  expect_error(DLY_LEVELS(PROV_TERR_STATE_LOC = "BCD", hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")))
-  expect_error(DLY_LEVELS(PROV_TERR_STATE_LOC = c("AB", "BCD"), hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")))
+test_that("hy_daily_levels produces an error when a province is not specified correctly", {
+  expect_error(hy_daily_levels(PROV_TERR_STATE_LOC = "BCD", hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")))
+  expect_error(hy_daily_levels(PROV_TERR_STATE_LOC = c("AB", "BCD"), hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")))
 })
 
 ## Too much data
-# test_that("DLY_LEVELS gather data when no arguments are supplied",{
-#  expect_true(nrow(DLY_LEVELS(hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat"))) >= 1)
+# test_that("hy_daily_levels gather data when no arguments are supplied",{
[…]
