+++ tests/testthat/test_download_realtime.R
@@ -1,11 +1,11 @@
-context("Testing download_realtime functions")
+context("Testing realtime functions")
 
- test_that("download_realtime_ws returns the correct data header", {
+ test_that("realtime_ws returns the correct data header", {
   skip_on_cran()
   skip_on_travis()
   token_out <- get_ws_token(username = Sys.getenv("WS_USRNM"), password = Sys.getenv("WS_PWD"))
 
-  ws_test <- download_realtime_ws(STATION_NUMBER = "08MF005",
+  ws_test <- realtime_ws(STATION_NUMBER = "08MF005",
                                parameters = c(46), ## Water level and temperature
                                start_date = Sys.Date(),
                                end_date = Sys.Date(),
@@ -20,26 +20,26 @@ context("Testing download_realtime functions")
  })
  
  
- test_that("download_realtime_dd returns the correct data header", {
+ test_that("realtime_dd returns the correct data header", {
    skip_on_cran()
    expect_identical(
-     colnames(download_realtime_dd(STATION_NUMBER = "08MF005", PROV_TERR_STATE_LOC = "BC")),
+     colnames(realtime_dd(STATION_NUMBER = "08MF005", PROV_TERR_STATE_LOC = "BC")),
      c("STATION_NUMBER", "PROV_TERR_STATE_LOC", "Date", "Parameter", "Value", "Grade", "Symbol", "Code")
   )
 })
 
-test_that("download_realtime_dd can download stations from multiple provinces using PROV_TERR_STATE_LOC", {
+test_that("realtime_dd can download stations from multiple provinces using PROV_TERR_STATE_LOC", {
   skip_on_cran()
-  download_realtime_dd(PROV_TERR_STATE_LOC = c("QC", "PE"))
+  realtime_dd(PROV_TERR_STATE_LOC = c("QC", "PE"))
 })
 
 
-test_that("download_realtime_dd can download stations from multiple provinces using STATION_NUMBER", {
+test_that("realtime_dd can download stations from multiple provinces using STATION_NUMBER", {
   skip_on_cran()
-  expect_error(download_realtime_dd(STATION_NUMBER = c("01CD005", "08MF005")), regexp = NA)
+  expect_error(realtime_dd(STATION_NUMBER = c("01CD005", "08MF005")), regexp = NA)
 })
 
 test_that("When STATION_NUMBER is ALL there is an error", {
   skip_on_cran()
-  expect_error(download_realtime_dd(STATION_NUMBER = "ALL"))
+  expect_error(realtime_dd(STATION_NUMBER = "ALL"))
 })

[…]
