### 2017-10-24 b7d8a24
change functions name per NEWS.md. 1st step to #45
