+++ R/SED_SAMPLES_PSD.R
@@ -12,11 +12,11 @@
 
 #' Extract instantaneous sediment sample information from the HYDAT database
 #'
-#' Provides wrapper to turn the SED_SAMPLES_PSD table in HYDAT into a tidy data frame.  \code{STATION_NUMBER} and
-#' \code{PROV_TERR_STATE_LOC} can both be supplied. If both are omitted all values from the \code{STATIONS} table are returned.
-#' That is a large vector for \code{SED_SAMPLES_PSD}.
+#' Provides wrapper to turn the hy_sed_samples_psd table in HYDAT into a tidy data frame.  \code{STATION_NUMBER} and
+#' \code{PROV_TERR_STATE_LOC} can both be supplied. If both are omitted all values from the \code{hy_stations} table are returned.
+#' That is a large vector for \code{hy_sed_samples_psd}.
 #'
-#' @inheritParams STATIONS
+#' @inheritParams hy_stations
 #' @param start_date Leave blank if all dates are required. Date format needs to be in YYYY-MM-DD. Date is inclusive.
 #' @param end_date Leave blank if all dates are required. Date format needs to be in YYYY-MM-DD. Date is inclusive.
 #'
@@ -25,7 +25,7 @@
 #' @examples
 #' \donttest{
 #'
-#' SED_SAMPLES_PSD(STATION_NUMBER = "01CA004")
+#' hy_sed_samples_psd(STATION_NUMBER = "01CA004")
 #'           }
 #'
 #' @family HYDAT functions
@@ -34,7 +34,7 @@
 
 
 
-SED_SAMPLES_PSD <- function(STATION_NUMBER = NULL,
+hy_sed_samples_psd <- function(STATION_NUMBER = NULL,
                             hydat_path = paste0(rappdirs::user_data_dir(),"\\Hydat.sqlite3"), 
                             PROV_TERR_STATE_LOC = NULL, start_date ="ALL", end_date = "ALL") {
   
