+++ R/MONTHLY_LEVELS.R
@@ -12,11 +12,11 @@
 
 #' Extract monthly levels information from the HYDAT database
 #'
-#' Provides wrapper to turn the MONTHLY_LEVELS table in HYDAT into a tidy data frame.  \code{STATION_NUMBER} and
-#'   \code{PROV_TERR_STATE_LOC} can both be supplied. If both are omitted all values from the \code{STATIONS} table are returned.
-#'   That is a large vector for \code{MONTHLY_LEVELS}.
+#' Provides wrapper to turn the hy_monthly_levels table in HYDAT into a tidy data frame.  \code{STATION_NUMBER} and
+#'   \code{PROV_TERR_STATE_LOC} can both be supplied. If both are omitted all values from the \code{hy_stations} table are returned.
+#'   That is a large vector for \code{hy_monthly_levels}.
 #'
-#' @inheritParams STATIONS
+#' @inheritParams hy_stations
 #' @param start_date Leave blank if all dates are required. Date format needs to be in YYYY-MM-DD. Date is inclusive.
 #' @param end_date Leave blank if all dates are required. Date format needs to be in YYYY-MM-DD. Date is inclusive.
 #'
@@ -25,10 +25,10 @@
 #'
 #' @examples
 #' \donttest{
-#' MONTHLY_LEVELS(STATION_NUMBER = c("02JE013","08MF005"), 
+#' hy_monthly_levels(STATION_NUMBER = c("02JE013","08MF005"), 
 #'   start_date = "1996-01-01", end_date = "2000-01-01")
 #'
-#' MONTHLY_LEVELS(PROV_TERR_STATE_LOC = "PE", hydat_path = "H:/Hydat.sqlite3")
+#' hy_monthly_levels(PROV_TERR_STATE_LOC = "PE", hydat_path = "H:/Hydat.sqlite3")
 #'           }
 #' @family HYDAT functions
 #' @source HYDAT
@@ -36,11 +36,11 @@
 
 
 
-MONTHLY_LEVELS <- function(STATION_NUMBER = NULL,
+hy_monthly_levels <- function(STATION_NUMBER = NULL,
                            hydat_path = paste0(rappdirs::user_data_dir(),"\\Hydat.sqlite3"),
                            PROV_TERR_STATE_LOC = NULL, start_date ="ALL", end_date = "ALL") {
   if (!is.null(STATION_NUMBER) && STATION_NUMBER == "ALL") {
-    stop("Deprecated behaviour.Omit the STATION_NUMBER = \"ALL\" argument. See ?MONTHLY_LEVELS for examples.")
+    stop("Deprecated behaviour.Omit the STATION_NUMBER = \"ALL\" argument. See ?hy_monthly_levels for examples.")
   }
   
   ## Check if hydat is present
