+++ R/SED_DLY_LOADS.R
@@ -12,11 +12,11 @@
 
 #' Extract daily sediment load information from the HYDAT database
 #'
-#' Provides wrapper to turn the SED_DLY_LOADS table in HYDAT into a tidy data frame.  \code{STATION_NUMBER} and
-#' \code{PROV_TERR_STATE_LOC} can both be supplied. If both are omitted all values from the \code{STATIONS} table are returned.
-#' That is a large vector for \code{SED_DLY_LOADS}.
+#' Provides wrapper to turn the hy_sed_daily_loads table in HYDAT into a tidy data frame.  \code{STATION_NUMBER} and
+#' \code{PROV_TERR_STATE_LOC} can both be supplied. If both are omitted all values from the \code{hy_stations} table are returned.
+#' That is a large vector for \code{hy_sed_daily_loads}.
 #'
-#' @inheritParams STATIONS
+#' @inheritParams hy_stations
 #' @param start_date Leave blank if all dates are required. Date format needs to be in YYYY-MM-DD. Date is inclusive.
 #' @param end_date Leave blank if all dates are required. Date format needs to be in YYYY-MM-DD. Date is inclusive.
 #'
@@ -24,7 +24,7 @@
 #'
 #' @examples
 #' \donttest{
-#' SED_DLY_LOADS(PROV_TERR_STATE_LOC = "PE", hydat_path = "H:/Hydat.sqlite3")
+#' hy_sed_daily_loads(PROV_TERR_STATE_LOC = "PE", hydat_path = "H:/Hydat.sqlite3")
 #'           }
 #'
 #' @family HYDAT functions
@@ -33,7 +33,7 @@
 
 
 
-SED_DLY_LOADS <- function(STATION_NUMBER = NULL, 
+hy_sed_daily_loads <- function(STATION_NUMBER = NULL, 
                           hydat_path = paste0(rappdirs::user_data_dir(),"\\Hydat.sqlite3"), 
                           PROV_TERR_STATE_LOC = NULL, start_date ="ALL", end_date = "ALL") {
   

[…]
