### 2017-10-25 506f74b
update test function and argument names. #45
