+++ tests/testthat/test_DLY_FLOWS.R
@@ -2,36 +2,72 @@ context("Testing hy_daily_flows")
 
 test_that("hy_daily_flows accepts single and multiple station arguments", {
   stns <- "08MF005"
-  expect_identical(unique(hy_daily_flows(STATION_NUMBER = stns, 
-                                    hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat"))$STATION_NUMBER), stns)
-  expect_identical(length(unique(hy_daily_flows(STATION_NUMBER = c("08MF005", "05AA008"), hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat"))$STATION_NUMBER)), length(c("08NM083", "08NE102")))
+  expect_identical(unique(
+    hy_daily_flows(
+      station_number = stns,
+      hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")
+    )$STATION_NUMBER
+  ), stns)
+  expect_identical(length(unique(
+    hy_daily_flows(
+      station_number = c("08MF005", "05AA008"),
+      hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")
+    )$STATION_NUMBER
+  )), length(c("08NM083", "08NE102")))
 })
 
 
-test_that("hy_daily_flows accepts single and multiple province arguments", {
-  expect_true(nrow(hy_daily_flows(PROV_TERR_STATE_LOC = "BC", hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat"))) >= 1)
-  expect_true(nrow(hy_daily_flows(PROV_TERR_STATE_LOC = c("BC", "YT"), hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat"))) >= 1)
-})
+test_that("hy_daily_flows accepts single and multiple province arguments",
+          {
+            expect_true(nrow(
+              hy_daily_flows(
+                prov_terr_state_loc = "BC",
+                hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")
+              )
+            ) >= 1)
+            expect_true(nrow(
+              hy_daily_flows(
+                prov_terr_state_loc = c("BC", "YT"),
+                hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")
+              )
+            ) >= 1)
+          })
 
-test_that("hy_daily_flows produces an error when a province is not specified correctly", {
-  expect_error(hy_daily_flows(PROV_TERR_STATE_LOC = "BCD", hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")))
-  expect_error(hy_daily_flows(PROV_TERR_STATE_LOC = c("YT", "BCD"), hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")))
-})
+test_that("hy_daily_flows produces an error when a province is not specified correctly",
+          {
+            expect_error(hy_daily_flows(
+              prov_terr_state_loc = "BCD",
+              hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")
+            ))
+            expect_error(hy_daily_flows(
+              prov_terr_state_loc = c("YT", "BCD"),
+              hydat_path = system.file("test_db/tinyhydat.sqlite3", package = "tidyhydat")
+            ))
+          })
 
 ## Too much data
 # test_that("hy_daily_flows gather data when no arguments are supplied",{
[…]
