+++ R/download.R
@@ -17,9 +17,9 @@
 #' downloading data collected at the highest resolution. In instances where data is not available at high (hourly or higher) resolution
 #' daily averages are used. Currently, if a station does not exist or is not found, no data is returned.
 #'
-#' @param STATION_NUMBER Water Survey of Canada station number. If this argument is omitted from the function call, the value of \code{PROV_TERR_STATE_LOC}
+#' @param station_number Water Survey of Canada station number. If this argument is omitted from the function call, the value of \code{prov_terr_state_loc}
 #' is returned.
-#' @param PROV_TERR_STATE_LOC Province, state or territory. If this argument is omitted from the function call, the value of \code{STATION_NUMBER}
+#' @param prov_terr_state_loc Province, state or territory. If this argument is omitted from the function call, the value of \code{station_number}
 #' is returned.
 #'
 #' @return A tibble of water flow and level values. Time is return as UTC for consistency.
@@ -27,30 +27,30 @@
 #'
 #' @examples
 #' ## Download from multiple provinces
-#' realtime_dd(STATION_NUMBER=c("01CD005","08MF005"))
+#' realtime_dd(station_number=c("01CD005","08MF005"))
 #'
 #' # To download all stations in Prince Edward Island:
-#' realtime_dd(PROV_TERR_STATE_LOC = "PE")
+#' realtime_dd(prov_terr_state_loc = "PE")
 #' 
 #' @family realtime functions
 #' @export
-realtime_dd <- function(STATION_NUMBER = NULL, PROV_TERR_STATE_LOC) {
+realtime_dd <- function(station_number = NULL, prov_terr_state_loc) {
 
   ## TODO: HAve a warning message if not internet connection exists
-  if (!is.null(STATION_NUMBER) && STATION_NUMBER == "ALL") {
-    stop("Deprecated behaviour.Omit the STATION_NUMBER = \"ALL\" argument. See ?realtime_dd for examples.")
+  if (!is.null(station_number) && station_number == "ALL") {
+    stop("Deprecated behaviour.Omit the station_number = \"ALL\" argument. See ?realtime_dd for examples.")
   }
 
 
-  if (!is.null(STATION_NUMBER)) {
-    stns <- STATION_NUMBER
+  if (!is.null(station_number)) {
+    stns <- station_number
     choose_df <- realtime_stations()
     choose_df <- dplyr::filter(choose_df, STATION_NUMBER %in% stns)
-    choose_df <- dplyr::select(choose_df, STATION_NUMBER, PROV_TERR_STATE_LOC)
+    choose_df <- dplyr::select(choose_df, STATION_NUMBER, prov_terr_state_loc)
   }
 
-  if (is.null(STATION_NUMBER)) {
-    choose_df <- realtime_stations(PROV_TERR_STATE_LOC = PROV_TERR_STATE_LOC)
+  if (is.null(station_number)) {
+    choose_df <- realtime_stations(prov_terr_state_loc = prov_terr_state_loc)
     choose_df <- dplyr::select(choose_df, STATION_NUMBER, PROV_TERR_STATE_LOC)
   }
 
@@ -189,22 +189,22 @@ realtime_dd <- function(STATION_NUMBER = NULL, PROV_TERR_STATE_LOC) {
 #' An up to date dataframe of all stations in the Realtime Water Survey of Canada 
 #'   hydrometric network operated by Environment and Climate Change Canada
 #'
-#' @param PROV_TERR_STATE_LOC Province/State/Territory or Location. See examples for list of available options. 
+#' @param prov_terr_state_loc Province/State/Territory or Location. See examples for list of available options. 
 #'   realtime_stations() for all stations.
 #'
 #' @family realtime functions
[…]
