+++ R/SED_DLY_SUSCON.R
@@ -12,8 +12,8 @@
 
 #' Extract daily suspended sediment concentration information from the HYDAT database
 #'
-#' Provides wrapper to turn the hy_sed_daily_suscon table in HYDAT into a tidy data frame.  \code{STATION_NUMBER} and
-#' \code{PROV_TERR_STATE_LOC} can both be supplied. If both are omitted all values from the \code{hy_stations} table are returned.
+#' Provides wrapper to turn the hy_sed_daily_suscon table in HYDAT into a tidy data frame.  \code{station_number} and
+#' \code{prov_terr_state_loc} can both be supplied. If both are omitted all values from the \code{hy_stations} table are returned.
 #' That is a large vector for \code{hy_sed_daily_suscon}.
 #'
 #' @inheritParams hy_daily_flows
@@ -23,7 +23,7 @@
 #' @examples
 #' \donttest{
 #'
-#' hy_sed_daily_suscon(STATION_NUMBER = "01CE003")
+#' hy_sed_daily_suscon(station_number = "01CE003")
 #'
 #'           }
 #'
@@ -33,9 +33,9 @@
 
 
 
-hy_sed_daily_suscon <- function(STATION_NUMBER = NULL,
+hy_sed_daily_suscon <- function(station_number = NULL,
                            hydat_path = paste0(rappdirs::user_data_dir(),"\\Hydat.sqlite3"), 
-                           PROV_TERR_STATE_LOC = NULL, 
+                           prov_terr_state_loc = NULL, 
                            start_date ="ALL", end_date = "ALL", symbol_output = "code") {
   
   ## Check if hydat is present
@@ -76,7 +76,7 @@ hy_sed_daily_suscon <- function(STATION_NUMBER = NULL,
   on.exit(DBI::dbDisconnect(hydat_con))
 
   ## Determine which stations we are querying
-  stns <- station_choice(hydat_con, STATION_NUMBER, PROV_TERR_STATE_LOC)
+  stns <- station_choice(hydat_con, station_number, prov_terr_state_loc)
 
   ## Data manipulations
   sed_dly_suscon <- dplyr::tbl(hydat_con, "SED_DLY_SUSCON")

[…]
