### 2017-10-25 63fe52c
lower case internal hydat tables. Another for #45
