+++ R/hy_instant_annual_peaks.R
@@ -66,10 +66,10 @@ hy_annual_instant_peaks <- function(station_number = NULL,
     dplyr::collect()
 
   ## Add in english data type
-  aip <- dplyr::left_join(aip, tidyhydat::DATA_TYPES, by = c("DATA_TYPE"))
+  aip <- dplyr::left_join(aip, tidyhydat::data_types, by = c("DATA_TYPE"))
 
   ## Add in Symbol
-  aip <- dplyr::left_join(aip, tidyhydat::DATA_SYMBOLS, by = c("SYMBOL" = "SYMBOL_ID"))
+  aip <- dplyr::left_join(aip, tidyhydat::data_symbols, by = c("SYMBOL" = "SYMBOL_ID"))
 
   ## If a yearis supplied...
   if (start_year != "ALL" | end_year != "ALL") {
