+++ R/zzz.R
@@ -22,7 +22,7 @@
           "NO_DAYS",
           "variable",
           "temp",
-          "DATA_SYMBOLS",
+          "data_symbols",
           "FLOW_SYMBOL",
           "SYMBOL_EN",
           "Value",
@@ -43,7 +43,7 @@
           "GRADE",
           "SYMBOL",
           "DATA_TYPE",
-          "DATA_TYPES",
+          "data_types",
           "DATA_TYPE_EN",
           "MAX_DAY",
           "MAX_MONTH",
