+++ man/bcstations.Rd
@@ -1,17 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/data.R
-\docType{data}
-\name{bcstations}
-\alias{bcstations}
-\title{A tibble of BC stations and their names}
-\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2280 rows and 2 columns.}
-\source{
-HYDAT
-}
-\usage{
-bcstations
-}
-\description{
-A shorthand to avoid having always call \code{hy_stations}. Only up to date as of 2017-07-18.
-}
-\keyword{datasets}
