+++ data-raw/HYDAT_internal_data/process_internal_data.R
@@ -7,14 +7,13 @@ library(tidyverse)
 
 ## Load up hydat connection
 ## Read in database
-hydat_con <- DBI::dbConnect(RSQLite::SQLite(), Sys.getenv("hydat"))
+hydat_con <- DBI::dbConnect(RSQLite::SQLite(), paste0(rappdirs::user_data_dir(),"\\Hydat.sqlite3"))
 
-## bc stations
-bcstations <- tbl(hydat_con, "STATIONS") %>%
-  filter(PROV_TERR_STATE_LOC == "BC") %>%
-  select(STATION_NUMBER, STATION_NAME) %>%
-  collect()
-use_data(bcstations, overwrite = TRUE)
+## To add to list
+# allstations <- download_network() %>%
+#  anti_join(hy_stations(), by = c("station_number", "STATION_NAME")) %>%
+#  select(STATION_NUMBER, STATION_NAME) 
+#' A tibble of all Canadian Stations stations and their names.
 
 ## all stations
 allstations <- tbl(hydat_con, "STATIONS") %>%
@@ -23,13 +22,13 @@ allstations <- tbl(hydat_con, "STATIONS") %>%
 use_data(allstations, overwrite = TRUE)
 
 ## DATA_TYPES
-DATA_TYPES <- tbl(hydat_con, "DATA_TYPES") %>%
+data_types <- tbl(hydat_con, "DATA_TYPES") %>%
   collect()
-use_data(DATA_TYPES, overwrite = TRUE)
+use_data(data_types, overwrite = TRUE)
 
 ## DATA_SYMBOLS
-DATA_SYMBOLS <- tbl(hydat_con, "DATA_SYMBOLS") %>%
+data_symbols <- tbl(hydat_con, "DATA_SYMBOLS") %>%
   collect()
-use_data(DATA_SYMBOLS, overwrite = TRUE)
+use_data(data_symbols, overwrite = TRUE)
 
 DBI::dbDisconnect(hydat_con)
