+++ data-raw/HYDAT_internal_data/README.md
@@ -4,4 +4,4 @@
 
 - The file `inst/test_db/tinyhydat.sqlite3` is the first 100 rows (or less) of each table in the hydat database. This is used exclusively for testing purposes and is generated by `tinyhydat_proc.R`
 
-- All other internal data (`allstations`,`bcstations`,`DATA_TYPES`,`DATA_SYMBOLS`) are drawn from HYDAT and are provided under the ([Open Government License - Canada](http://open.canada.ca/en/open-government-licence-canada)). 
+- All other internal data (`allstations`,`data_types`,`data_symbols`) are drawn from HYDAT and are provided under the ([Open Government License - Canada](http://open.canada.ca/en/open-government-licence-canada)). 
