### 2017-10-25 f4ddaf6
Update docs with new filename. More #45
