+++ man/hy_sed_monthly_loads.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/SED_MONTHLY_LOADS.R
+% Please edit documentation in R/hy_sed_monthly_loads.R
 \name{hy_sed_monthly_loads}
 \alias{hy_sed_monthly_loads}
 \title{Extract monthly flows information from the HYDAT database}
