+++ man/hy_daily_levels.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/DLY_LEVELS.R
+% Please edit documentation in R/hy_daily_levels.R
 \name{hy_daily_levels}
 \alias{hy_daily_levels}
 \title{Extract daily levels information from the HYDAT database}
