+++ man/hy_sed_monthly_suscon.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/SED_MONTHLY_SUSCON.R
+% Please edit documentation in R/hy_sed_monthly_suscon.R
 \name{hy_sed_monthly_suscon}
 \alias{hy_sed_monthly_suscon}
 \title{Extract monthly flows information from the HYDAT database}
