+++ man/hy_monthly_flows.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/MONTHLY_FLOWS.R
+% Please edit documentation in R/hy_monthly_flows.R
 \name{hy_monthly_flows}
 \alias{hy_monthly_flows}
 \title{Extract monthly flows information from the HYDAT database}
