+++ man/hy_annual_instant_peaks.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/ANNUAL_INSTANT_PEAKS.R
+% Please edit documentation in R/hy_instant_annual_peaks.R
 \name{hy_annual_instant_peaks}
 \alias{hy_annual_instant_peaks}
 \title{Annual maximum/minimum instantaneous flows and water levels}
@@ -34,12 +34,12 @@ Provides wrapper to turn the hy_annual_instant_peaks table in HYDAT into a tidy 
 \code{prov_terr_state_loc} can both be supplied.
 }
 \examples{
-\donttest{
+\dontrun{
 ## Multiple stations province not specified
 hy_annual_instant_peaks(station_number = c("08NM083","08NE102"))
 
 ## Multiple province, station number not specified
-hy_annual_instant_peaks(prov_terr_state_loc = c("AB","YT"), hydat_path = "H:/Hydat.sqlite3")
+hy_annual_instant_peaks(prov_terr_state_loc = c("AB","YT"))
 }
 
 }
