+++ man/hy_stn_regulation.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/STN_REGULATION.R
+% Please edit documentation in R/hy_stn_regulation.R
 \name{hy_stn_regulation}
 \alias{hy_stn_regulation}
 \title{Extract station regulation from the HYDAT database}
