+++ man/hy_sed_daily_suscon.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/SED_DLY_SUSCON.R
+% Please edit documentation in R/hy_sed_daily_suscon.R
 \name{hy_sed_daily_suscon}
 \alias{hy_sed_daily_suscon}
 \title{Extract daily suspended sediment concentration information from the HYDAT database}
