+++ man/hy_monthly_levels.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/MONTHLY_LEVELS.R
+% Please edit documentation in R/hy_monthly_levels.R
 \name{hy_monthly_levels}
 \alias{hy_monthly_levels}
 \title{Extract monthly levels information from the HYDAT database}
