+++ man/hy_stn_data_range.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/STN_.R
+% Please edit documentation in R/hy_stn_.R
 \name{hy_stn_data_range}
 \alias{hy_stn_data_range}
 \title{hy_stn_data_range function}
