+++ man/hy_sed_samples_psd.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/SED_SAMPLES_PSD.R
+% Please edit documentation in R/hy_sed_samples_psd.R
 \name{hy_sed_samples_psd}
 \alias{hy_sed_samples_psd}
 \title{Extract instantaneous sediment sample information from the HYDAT database}
