+++ R/hy_sed_daily_loads.R
@@ -34,8 +34,12 @@
 
 
 hy_sed_daily_loads <- function(station_number = NULL, 
-                          hydat_path = paste0(rappdirs::user_data_dir(),"\\Hydat.sqlite3"), 
+                          hydat_path = NULL, 
                           prov_terr_state_loc = NULL, start_date ="ALL", end_date = "ALL") {
+  
+  if(is.null(hydat_path)){
+    hydat_path = paste0(rappdirs::user_data_dir(),"\\Hydat.sqlite3")
+  }
   
   ## Check if hydat is present
   if (!file.exists(hydat_path)){
