+++ R/hy_daily_levels.R
@@ -36,7 +36,7 @@
 
 
 hy_daily_levels <- function(station_number = NULL, 
-                       hydat_path = paste0(rappdirs::user_data_dir(),"\\Hydat.sqlite3"),
+                       hydat_path = NULL,
                        prov_terr_state_loc = NULL, 
                        start_date ="ALL", end_date = "ALL", symbol_output = "code") {
   if (!is.null(station_number) && station_number == "ALL") {
@@ -69,6 +69,10 @@ hy_daily_levels <- function(station_number = NULL,
     }
   }
 
+  if(is.null(hydat_path)){
+    hydat_path = paste0(rappdirs::user_data_dir(),"\\Hydat.sqlite3")
+  }
+  
   ## Check if hydat is present
   if (!file.exists(hydat_path)){
     stop(paste0("No Hydat.sqlite3 found at ",rappdirs::user_data_dir(),". Run download_hydat() to download the database."))
