+++ R/hy_sed_samples_psd.R
@@ -35,8 +35,12 @@
 
 
 hy_sed_samples_psd <- function(station_number = NULL,
-                            hydat_path = paste0(rappdirs::user_data_dir(),"\\Hydat.sqlite3"), 
+                            hydat_path = NULL, 
                             prov_terr_state_loc = NULL, start_date ="ALL", end_date = "ALL") {
+  
+  if(is.null(hydat_path)){
+    hydat_path = paste0(rappdirs::user_data_dir(),"\\Hydat.sqlite3")
+  }
   
   ## Check if hydat is present
   if (!file.exists(hydat_path)){
