+++ R/hy_annual_stats.R
@@ -36,13 +36,17 @@
 #' @export
 
 hy_annual_stats <- function(station_number =NULL, 
-                              hydat_path = paste0(rappdirs::user_data_dir(),"\\Hydat.sqlite3"), 
+                              hydat_path = NULL, 
                               prov_terr_state_loc=NULL,
                               start_year = "ALL", end_year = "ALL") {
   
   if (!is.null(station_number) && station_number == "ALL") {
     stop("Deprecated behaviour.Omit the station_number = 
          \"ALL\" argument. See ?hy_annual_stats for examples.")
+  }
+  
+  if(is.null(hydat_path)){
+    hydat_path = paste0(rappdirs::user_data_dir(),"\\Hydat.sqlite3")
   }
   
   ## Check if hydat is present
