### 2017-10-25 bca30ea
update hydat_path argument description
