+++ data-raw/HYDAT_internal_data/process_internal_data.R
@@ -5,21 +5,19 @@ library(tidyverse)
 ##param_id
 
 
+
+
+#' A tibble of all Canadian Stations stations and their names.
+allstations <- realtime_stations() %>%
+  bind_rows(hy_stations()) %>%
+  distinct(STATION_NUMBER, .keep_all = TRUE) %>%
+  select(STATION_NUMBER, STATION_NAME, PROV_TERR_STATE_LOC, LATITUDE, LONGITUDE) 
+
+use_data(allstations, overwrite = TRUE)
+
 ## Load up hydat connection
 ## Read in database
 hydat_con <- DBI::dbConnect(RSQLite::SQLite(), paste0(rappdirs::user_data_dir(),"\\Hydat.sqlite3"))
-
-## To add to list
-# allstations <- download_network() %>%
-#  anti_join(hy_stations(), by = c("station_number", "STATION_NAME")) %>%
-#  select(STATION_NUMBER, STATION_NAME) 
-#' A tibble of all Canadian Stations stations and their names.
-
-## all stations
-allstations <- tbl(hydat_con, "STATIONS") %>%
-  select(STATION_NUMBER, STATION_NAME, PROV_TERR_STATE_LOC) %>%
-  collect()
-use_data(allstations, overwrite = TRUE)
 
 ## DATA_TYPES
 data_types <- tbl(hydat_con, "DATA_TYPES") %>%
