### 2017-10-26 d45fc87
small bug resulting from #45. Fixed now.
