### 2017-10-26 422f76f
capitalize internal ws function
