### 2017-10-26 4aff53e
search functions now search all available data. Some tests and binding variable added. Closes #56
