### 2017-10-26 328822c
skipping search tests on cran and travis
