### 2017-10-26 fe4f70f
revised function help pages to include units and columns outputted. Closes #49
