+++ R/hy_sed_samples.R
@@ -21,12 +21,31 @@
 #' @param end_date Leave blank if all dates are required. Date format needs to be in YYYY-MM-DD. Date is inclusive.
 #'
 #' @return A tibble of instantaneous sediment samples data
+#' 
+#' @format A tibble with 19 variables:
+#' \describe{
+#'   \item{STATION_NUMBER}{Unique 7 digit Water Survey of Canada station number}
+#'   \item{SED_DATA_TYPE}{Contains the type of sampling method used in collecting sediment for a station}
+#'   \item{DATE}{Contains the time to the nearest minute of when the sample was taken}
+#'   \item{SAMPLE_REMARK_CODE}{Descriptive Sediment Sample Remark in English}
+#'   \item{TIME_SYMBOL}{An "E" symbol means the time is an estimate only}
+#'   \item{FLOW}{Contains the instantaneous discharge in cubic metres per second at the time the sample was taken}
+#'   \item{SYMBOL_EN}{Indicates a condition where the daily mean has a larger than expected error}
+#'   \item{SAMPLER_TYPE}{Contains the type of measurement device used to take the sample}
+#'   \item{SAMPLING_VERTICAL_LOCATION}{The location on the cross-section of the river 
+#'         at which the single sediment samples are collected. If one of the standard 
+#'         locations is not used the distance in meters will be shown}
+#'   \item{SAMPLING_VERTICAL_EN}{ndicates sample location relative to the 
+#'         regular measurement cross-section or the regular sampling site}
+#'   \item{TEMPERATURE}{Contains the instantaneous water temperature
+#'         in Celsius at the time the sample was taken}
+#'   \item{CONCENTRATION_EN}{Contains the instantaneous concentration sampled in milligrams per litre}
+#'   \item{SV_DEPTH2}{Depth 2 for split vertical depth integrating (m)}
+#' }
 #'
 #' @examples
-#' \donttest{
-#'
+#' \dontrun{
 #' hy_sed_samples(station_number = "01CA004")
-#'
 #'           }
 #'
 #' @family HYDAT functions
@@ -114,7 +133,7 @@ hy_sed_samples <- function(station_number = NULL,
   sed_samples <- dplyr::select(
     sed_samples, STATION_NUMBER, SED_DATA_TYPE_EN, DATE, SAMPLE_REMARK_EN, TIME_SYMBOL,
     FLOW, SYMBOL_EN, SAMPLER_TYPE, SAMPLING_VERTICAL_LOCATION, SAMPLING_VERTICAL_EN,
-    TEMPERATURE, CONCENTRATION, CONCENTRATION_EN:SV_DEPTH2
+    TEMPERATURE, CONCENTRATION, CONCENTRATION_EN, SV_DEPTH2
   )
   
   
