+++ R/hy_monthly_flows.R
@@ -20,16 +20,29 @@
 #' @param start_date Leave blank if all dates are required. Date format needs to be in YYYY-MM-DD. Date is inclusive.
 #' @param end_date Leave blank if all dates are required. Date format needs to be in YYYY-MM-DD. Date is inclusive.
 #'
-#' @return A tibble of monthly flows. This includes a \code{Date_occured} column which indicates the date of the \code{Sum_stat}. For MEAN and 
-#'   TOTAL this is not presented as those are not daily values.
+#' @return A tibble of monthly flows. 
+#'   
+#' @format A tibble with 8 variables:
+#' \describe{
+#'   \item{STATION_NUMBER}{Unique 7 digit Water Survey of Canada station number}
+#'   \item{YEAR}{Year of record.}
+#'   \item{MONTH}{Numeric month value}
+#'   \item{FULL_MONTH}{Logical value is there is full record from MONTH}
+#'   \item{NO_DAYS}{Number of days in that month}
+#'   \item{Sum_stat}{Summary statistic being used.} 
+#'   \item{Value}{Value of the measurement in m^3/s.}
+#'   \item{Date_occurred}{Observation date. Formatted as a Date class. MEAN is a annual summary 
+#'   and therefore has an NA value for Date.}
+#' }
 #'
 #' @examples
-#' \donttest{
+#' \dontrun{
 #' hy_monthly_flows(station_number = c("02JE013","08MF005"), 
 #'   start_date = "1996-01-01", end_date = "2000-01-01")
 #'
-#' hy_monthly_flows(prov_terr_state_loc = "PE", hydat_path = "H:/Hydat.sqlite3")
+#' hy_monthly_flows(prov_terr_state_loc = "PE")
 #'           }
+#'           
 #' @family HYDAT functions
 #' @source HYDAT
 #' @export
