### 2017-10-26 91cb6b5
Better @description in man pages. Addresses partial #48
