+++ R/hy_stn_regulation.R
@@ -14,8 +14,9 @@
 
 #' Extract station regulation from the HYDAT database
 #'
-#' Provides wrapper to turn the hy_stn_regulation table in HYDAT into a tidy data frame. \code{station_number} and
-#' \code{prov_terr_state_loc} can both be supplied. If both are omitted all values from the \code{hy_stations} table are returned
+#' Provides wrapper to turn the hy_stn_regulation table in HYDAT into a tidy data frame of station regulation. 
+#' \code{station_number} and \code{prov_terr_state_loc} can both be supplied. If both are omitted all values 
+#' from the \code{hy_stations} table are returned.
 #'
 #' @inheritParams hy_stations
 #'
