+++ R/hy_instant_annual_peaks.R
@@ -13,8 +13,8 @@
 
 #' Annual maximum/minimum instantaneous flows and water levels
 #'
-#' Provides wrapper to turn the hy_annual_instant_peaks table in HYDAT into a tidy data frame. \code{station_number} and
-#' \code{prov_terr_state_loc} can both be supplied. 
+#' Provides wrapper to turn the ANNUAL_INSTANT_PEAKS table in HYDAT into a tidy data frame of instantaneous flows and water levels. 
+#' \code{station_number} and \code{prov_terr_state_loc} can both be supplied. 
 #' 
 #' @inheritParams hy_stations
 #' @param start_year First year of the returned record
