+++ R/hy_monthly_levels.R
@@ -12,7 +12,7 @@
 
 #' Extract monthly levels information from the HYDAT database
 #'
-#' Provides wrapper to turn the hy_monthly_levels table in HYDAT into a tidy data frame.  \code{station_number} and
+#' Tidy data of monthly river or lake levels information from the DLY_LEVELS HYDAT table. \code{station_number} and
 #'   \code{prov_terr_state_loc} can both be supplied. If both are omitted all values from the \code{hy_stations} table are returned.
 #'   That is a large vector for \code{hy_monthly_levels}.
 #'
