### 2017-10-26 fc2b2d2
first draft at updated vignettes
