+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: tidyhydat
 Title: Extract and Tidy Canadian Hydrometric Data
-Version: 0.2.9
+Version: 0.3.0
 Authors@R: c(person("Sam", "Albers", email = "sam.albers@gov.bc.ca", role = c("aut", "cre"),
            ), person("David", "Hutchinson", email = "david.hutchinson@canada.ca", role = "ctb"),
            person("Province of British Columbia", role = "cph"))
@@ -23,7 +23,12 @@ Imports:
     httr (>= 1.3.1),
     rappdirs (>= 0.3.1)
 Suggests:
-    tidyverse,
+    purrr,
+    corrr,
+    igraph,
+    ggraph,
+    ggmap,
+    viridis,
     dbplyr,
     knitr,
     rmarkdown,
