### 2017-10-26 0091cd3
Update README and paper with new function names
