### 2017-10-26 aee1ee6
organized up the NEWS section a little
