### 2017-10-30 27cced4
trying to fix travis build failure with ggraph
