+++ DESCRIPTION
@@ -24,6 +24,7 @@ Imports:
     rappdirs (>= 0.3.1)
 Suggests:
     purrr,
+    ggplot2,
     corrr,
     igraph,
     ggraph,
