### 2017-10-30 9a5e671
trying to fix travis build failure with ggraph
