+++ DESCRIPTION
@@ -23,8 +23,8 @@ Imports:
     httr (>= 1.3.1),
     rappdirs (>= 0.3.1)
 Suggests:
-    purrr,
     ggplot2,
+    purrr,
     corrr,
     igraph,
     ggraph,
