### 2017-10-31 43e3740
fixed data frame name in stn_data_coll
