### 2017-10-31 4e5635d
removed igraph vignette and dependencies
