+++ DESCRIPTION
@@ -24,12 +24,6 @@ Imports:
     rappdirs (>= 0.3.1)
 Suggests:
     ggplot2,
-    purrr,
-    corrr,
-    igraph,
-    ggraph,
-    ggmap,
-    viridis,
     dbplyr,
     knitr,
     rmarkdown,
