+++ .travis.yml
@@ -1,11 +1,8 @@
 # R for travis: see documentation at https://docs.travis-ci.com/user/languages/r
 
 language: R
-sudo: true
+sudo: false
 cache: packages
-
-apt_packages:
-  - libudunits2-dev
 
 r_build_args: "--no-build-vignettes"
 r_check_args: "--no-build-vignettes --as-cran --no-examples"
