### 2017-10-31 988617e
2nd draft of vignettes
