+++ inst/doc/tidyhydat_example_analysis.Rmd
@@ -1,0 +1,181 @@
+---
+title: "Two examples of using tidyhydat"
+author: "Sam Albers"
+date: "`r Sys.Date()`"
+output:
+  html_vignette:
+     keep_md: true
+vignette: >
+  %\VignetteIndexEntry{tidyhydat}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r setup, include=FALSE}
+knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE, fig.width=7, fig.height=7)
+```
+
+# Introduction
+
+The real power of `tidyhydat` comes with its ability to efficiently import data into R which then can be used in conjunction with R's vast array of packages. `tidyhydat` gets hydrometric data into R and allows the user to start conducting data analysis immediately. This vignette is designed to take a user through two short examples of hydrological visualization and analysis and will therefore use additional tools beyond simply the functions in `tidyhydat`. 
+
+## Familiarity with the tidyverse
+`tidyhydat` is designed to exist as a "tidy tools" that is complementary to the `tidyverse` suite of packages. This vignette assumes some familiarity with basic tidy tools particularly `dplyr`, `tidyr`, `ggplot2`, `lubridate` and `purrr`. The [R for Data Science](http://r4ds.had.co.nz/) book by Garrett Grolemund and Hadley Wickham is an enormously useful book both for general R use and for `tidyhydat`. The following is a list of useful links to begin learning particular aspect of each `tidyverse` package:
+
+ - `dplyr`: http://dplyr.tidyverse.org/
+ - `tidyr`: http://tidyr.tidyverse.org/
+ - `ggplot2`: http://ggplot2.tidyverse.org/
+ - `purrr`: http://purrr.tidyverse.org/
+ - `lubridate`: http://lubridate.tidyverse.org/
+
+# Basic Data Extraction and Plotting
+
+## Objective
+Use `tidyhydat` to find the longest hydrometric network in Canada then use `ggplot2` to create a visualization of the discharge data. 
+
+In this section, we will be using `dplyr`, a data manipulation package, `ggplot2`, an extremely simple yet powerful data visualization tool, `lubridate` a package that aids significant in working with date and times in R and of course `tidyhydat`. Take a deep breathe. Though this might seem like an overwhelming onslaught of information on packages, each of these packages will save you considerable time with relatively minimal learning investment.
+```{r pkg_load_1}
+library(tidyhydat)
+library(dplyr)
+library(ggplot2)
+library(lubridate)
+```
+
+Your first step is to download the HYDAT database which `tidyhydat` facilitate for you:
+```{r dl_hy, eval=FALSE}
+download_hydat()
+```
+This should be the only instance where you will need to interact directly with HYDAT. Each `tidyhydat` function prefixed with `hy` will automatically know where to look for HYDAT saving you the trouble. When using `tidyhydat`, often your first task is to find the station(s) that you are interested in. Because we are interested in the longest record we can extract that information with `hy_stn_data_range()` and then feed that information to `hy_daily_flows()` like this: 
+
+```{r, eval= FALSE, warning=FALSE, message=FALSE}
+longest_record_data <- hy_stn_data_range() %>%
+  filter(RECORD_LENGTH == max(RECORD_LENGTH)) %>%
+  pull(STATION_NUMBER) %>%
+  hy_daily_flows()
+```
+Let's break this down line by line to understand how `tidyhydat` uses tidy tools. First we are interested in getting data on record length:
+```{r data_range}
+hy_stn_data_range()
+```
[…]
