### 2017-10-31 836740d
replace rappdirs to hy_dir wrapper
