+++ R/hy_annual_stats.R
@@ -59,12 +59,12 @@ hy_annual_stats <- function(station_number =NULL,
   }
   
   if(is.null(hydat_path)){
-    hydat_path = paste0(rappdirs::user_data_dir(),"\\Hydat.sqlite3")
+    hydat_path = paste0(hy_dir(),"\\Hydat.sqlite3")
   }
   
   ## Check if hydat is present
   if (!file.exists(hydat_path)){
-    stop(paste0("No Hydat.sqlite3 found at ",rappdirs::user_data_dir(),". Run download_hydat() to download the database."))
+    stop(paste0("No Hydat.sqlite3 found at ",hy_dir(),". Run download_hydat() to download the database."))
   }
 
   ## Read in database
