### 2017-10-31 68d7b42
create hy_dir function
