+++ NAMESPACE
@@ -7,6 +7,7 @@ export(hy_annual_stats)
 export(hy_daily_flows)
 export(hy_daily_levels)
 export(hy_datum_list)
+export(hy_dir)
 export(hy_monthly_flows)
 export(hy_monthly_levels)
 export(hy_reg_office_list)
