### 2017-10-31 7a47b9e
update NEWS
