### 2017-10-31 efa3e06
update README. Closes #46
