+++ README.md
@@ -8,7 +8,7 @@ tidyhydat <img src="img/tidyhydat.png" align="right" />
 Project Status
 --------------
 
-This package is under active development. The package is currently undergoing a review by [rOpenSci](https://ropensci.org/). You can see the progress of this review here: <https://github.com/ropensci/onboarding/issues/152>. As a result breaking changes to `tidyhydat` functions will be made. To install the most recent version prior to these changes you can use this command:
+This package is under active development. The package is currently undergoing a review by [rOpenSci](https://ropensci.org/). You can see the progress of this review here: <https://github.com/ropensci/onboarding/issues/152>. As a result, breaking changes to `tidyhydat` functions will be made. To install the most recent version prior to these changes you can use this command:
 
 ``` r
 devtools::install_github("bcgov/tidyhydat@v0.2.9")
@@ -19,26 +19,24 @@ However, you should strongly consider migrating to the updated `tidyhydat` once 
 What does it do?
 ----------------
 
-Here is a list of what `tidyhydat` does:
+Here is a summary of what `tidyhydat` does:
 
--   Provide function that access each table in the HYDAT database and return tidy data.
--   Keep functions as low-level as possible. For example, for daily flows, the `DLY_FLOWS()` function queries the database, *tidies* the data and returns the data.
--   Provide functions that access Environment and Climate Change Canada's real-time hydrometric data source.
--   Provide functions that search full station lists and aid in generating station vectors
-
-A more thorough vignette outlining the full functionality of `tidyhydat` is outlined [here](https://github.com/bcgov/tidyhydat/blob/master/vignettes/tidyhydat.Rmd)
+-   Provides functions (`hy_*`) that access hydrometric data from the HYDAT database, a national archive of Canadian hydrometric data and return tidy data.
+-   Provides functions (`realtime_*`) that access Environment and Climate Change Canada's real-time hydrometric data source.
+-   Provides functions (`search_*`) that can search through the approximately 7000 stations in the database and aid in generating station vectors
+-   Keep functions as simple as possible. For example, for daily flows, the `hy_daily_flows()` function queries the database, *tidies* the data and returns a [tibble](http://tibble.tidyverse.org/) of daily flows.
 
 Installation
 ------------
 
-To install the `tidyhydat` package, you need to install the `devtools` package then the `tidyhydat` package
+To install the `tidyhydat` package, you need to install the `remotes` package then the `tidyhydat` package
 
 ``` r
 install.packages("remotes")
 remotes::install_github("bcgov/tidyhydat")
 ```
 
-Then to load the package you need to using the `library()` function. When you install `tidyhydat`, several other packages will be installed as well. One of those packages, `dplyr`, is useful for data manipulations and is used regularly here. Even though `dplyr` is installed alongside `tidyhydat`, it is helpful to load it by itself as there are many useful functions contained within `dplyr`. A helpful `dplyr` tutorial can be found [here](https://cran.r-project.org/web/packages/dplyr/vignettes/dplyr.html).
[…]
