### 2017-10-31 7fb2782
move renviron credentials into the function
