+++ man/tidyhydat.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/tidyhydat.R
+% Please edit documentation in R/tidyhydat-package.R
 \docType{package}
 \name{tidyhydat}
 \alias{tidyhydat}
