### 2017-10-31 5be9209
.onAttach() checks if HYDAT is downloaded on package load
