+++ R/zzz.R
@@ -102,3 +102,25 @@
       )
     invisible()
   }
+
+# Won't pass cmd check with this uncommented
+.onAttach <- function(libname, pkgname) {
+
+  if(!file.exists(paste0(hy_dir(),"\\Hydat.sqlite3"))){
+    packageStartupMessage("Tidyhydat requires HYDAT which has not yet been downloaded. Download using download_hydat()")
+  }
+#  
+#  ## Check the date of current HYDAT
+#  base_url <- "http://collaboration.cmc.ec.gc.ca/cmc/hydrometrics/www/"
+#  x <- httr::GET(base_url)
+#  httr::stop_for_status(x)
+#  new_hydat <- substr(gsub(
+#    "^.*\\Hydat_sqlite3_", "",
+#    httr::content(x, "text")
+#  ), 1, 8)
+#
+#  # Do we need to download a new version?
+#  if (!(as.Date(new_hydat, "%Y%m%d") == as.Date(hy_version()$Date))) {
+#    packageStartupMessage(paste0("A new version of HYDAT was release on ", as.Date(new_hydat, "%Y%m%d"),". Use download_hydat() to get the new version."))
+#  }
+}
