### 2017-10-31 2b7c4d1
update NEWS
