### 2017-10-31 82d9659
removed typos (i hope) from function documentation. Closes #48
