### 2017-10-31 285e249
update vignettes
