+++ inst/doc/tidyhydat_an_introduction.Rmd
@@ -20,46 +20,55 @@ vignette: >
 ```{r options, include=FALSE}
 knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE, fig.width = 8, fig.height = 12)
 ```
-
-```{r packages, warning=FALSE, message=FALSE, echo = FALSE}
+## Package loading
+```{r packages, warning=FALSE, message=FALSE, echo = TRUE}
 library(tidyhydat)
-library(dbplyr)
+library(dplyr)
 ```
 
 # `tidyhydat` package
 This vignette will outline a few key options that will hopefully make `tidyhydat` useful.  
 
-### HYDAT download
-To use most of the `tidyhydat` package you will need to download a version of the HYDAT database, Environment and Climate Change Canada's comprehensive database of historical hydrometric data then tell R where to find the database. `tidyhydat` does all this for you via:
+## HYDAT download
+To use many of the functions in the `tidyhydat` package you will need to download a version of the HYDAT database, Environment and Climate Change Canada's database of historical hydrometric data then tell R where to find the database. Conveniently `tidyhydat` does all this for you via:
 ```{r, eval=FALSE}
 download_hydat()
 ```
-This downloads the most recent version of HYDAT and then saves it in a location on your computer where `tidyhydat`'s function will look for it. Do be patient though as this takes a long time! To see where HYDAT was saved you can run `rappdirs::site_data_dir()`.
+This downloads the most recent version of HYDAT and then saves it in a location on your computer where `tidyhydat`'s function will look for it. Do be patient though as this takes a long time! To see where HYDAT was saved you can run `hy_dir()`. Now that you have HYDAT downloaded and ready to go, you are all set to begin some hydrologic analysis.
+  
+## Usage
+Most functions in `tidyhydat` follow a common argument structure. We will use the `hy_daily_flows()` function for the following examples though the same approach applies to most functions in the package (See `ls("package:tidyhydat")` for a list of exported objects). Much of the functionality of `tidyhydat` originates with the choice of hydrometric stations that you are interested in. A user will often find themselves creating vectors of station numbers. There are several ways to do this. 
 
-### Usage
-All functions in `tidyhydat` follow a common argument structure which can be illustrated with the `DLY_FLOWS()` function. For these functions, you can supply either the `station_number` and the `prov_terr_state_loc` arguments. If you would like to query the database for two or more stations you could vectorize `station_number`:
-```{r warning=FALSE, warning=FALSE, message=FALSE, eval=FALSE}
-hy_daily_flows(station_number = c("08LA001","08NL071"))
+The simplest case is if you would like to extract only station. You can supply this directly to the `station_number` argument:
+```{r example1, warning=FALSE}
+hy_daily_flows(station_number = "08LA001")
 ```
 
-If instead you would like to extract flows for all stations from a jurisdiction, you can simply specific the province:
-```{r warning=FALSE, warning=FALSE, message=FALSE, eval=FALSE}
-hy_daily_flows(prov_terr_state_loc = "PE")
+Another method is to use `hy_stations()` to generate your vector which is then given the `station_number` argument. For example, we could take a subset for only those active stations within Prince Edward Island (Province code:PE) and then create vector for `hy_daily_flows()`:
+
+```{r example2, warning=FALSE}
+PEI_stns <- hy_stations() %>%
+  filter(HYD_STATUS == "ACTIVE") %>%
+  filter(PROV_TERR_STATE_LOC == "PE") %>%
+  pull(STATION_NUMBER)
+
+PEI_stns
+
+hy_daily_flows(station_number = PEI_stns)
+```
+
[…]
