### 2017-11-01 d7395fc
remove = from object assignment as per goodpractices
