+++ R/hy_stn_regulation.R
@@ -49,7 +49,7 @@ hy_stn_regulation <- function(station_number = NULL,
                            prov_terr_state_loc = NULL) {
   
   if(is.null(hydat_path)){
-    hydat_path = paste0(hy_dir(),"\\Hydat.sqlite3")
+    hydat_path <- paste0(hy_dir(),"\\Hydat.sqlite3")
   }
   
   ## Check if hydat is present
