+++ R/hy_daily_levels.R
@@ -78,7 +78,7 @@ hy_daily_levels <- function(station_number = NULL,
   }
 
   if(is.null(hydat_path)){
-    hydat_path = paste0(hy_dir(),"\\Hydat.sqlite3")
+    hydat_path <- paste0(hy_dir(),"\\Hydat.sqlite3")
   }
   
   ## Check if hydat is present
