+++ man/allstations.Rd
@@ -13,7 +13,7 @@
   \item{LONGITUDE}{East-West Coordinates of the gauging station in decimal degrees}
 }}
 \source{
-HYDAT, Meterological Service of Canada datamart
+HYDAT, Meteorological Service of Canada datamart
 }
 \usage{
 allstations
