### 2017-11-01 0b3e0aa
added some additional tests
