### 2017-11-01 939f9b1
fixed hydat paths to be platform independent. Closes #57
