+++ R/hy_stations.R
@@ -73,7 +73,7 @@ hy_stations <- function(station_number = NULL,
   }
 
   if(is.null(hydat_path)){
-    hydat_path <- paste0(hy_dir(),"\\Hydat.sqlite3")
+    hydat_path <- file.path(hy_dir(),"Hydat.sqlite3")
   }
   
   ## Check if hydat is present
