+++ R/hy_annual_stats.R
@@ -59,7 +59,7 @@ hy_annual_stats <- function(station_number =NULL,
   }
   
   if(is.null(hydat_path)){
-    hydat_path <- paste0(hy_dir(),"\\Hydat.sqlite3")
+    hydat_path <- file.path(hy_dir(),"Hydat.sqlite3")
   }
   
   ## Check if hydat is present
