+++ R/hy_sed_monthly_suscon.R
@@ -53,7 +53,7 @@ hy_sed_monthly_suscon <- function(station_number = NULL,
   }
   
   if(is.null(hydat_path)){
-    hydat_path <- paste0(hy_dir(),"\\Hydat.sqlite3")
+    hydat_path <- file.path(hy_dir(),"Hydat.sqlite3")
   }
   
   ## Check if hydat is present
