+++ data-raw/HYDAT_internal_data/process_internal_data.R
@@ -17,7 +17,7 @@ use_data(allstations, overwrite = TRUE)
 
 ## Load up hydat connection
 ## Read in database
-hydat_con <- DBI::dbConnect(RSQLite::SQLite(), paste0(hy_dir(),"\\Hydat.sqlite3"))
+hydat_con <- DBI::dbConnect(RSQLite::SQLite(), file.path(hy_dir(),"Hydat.sqlite3"))
 
 ## DATA_TYPES
 hy_data_types <- tbl(hydat_con, "DATA_TYPES") %>%
