+++ R/hy_monthly_flows.R
@@ -57,7 +57,7 @@ hy_monthly_flows <- function(station_number = NULL,
   }
   
   if(is.null(hydat_path)){
-    hydat_path <- paste0(hy_dir(),"\\Hydat.sqlite3")
+    hydat_path <- file.path(hy_dir(),"Hydat.sqlite3")
   }
   
   ## Check if hydat is present
