### 2017-11-01 a2ac9e1
misc win build issues
