### 2017-11-01 bcb1f25
slight typo in NEWS
