### 2017-11-03 #152 @sckott
Ran through some examples and such - no comments from me.

@lawinslow and @ldecicco-USGS - are you happy with the changes?
