### 2017-11-07 673f05a
additional win builder notes
