### 2017-11-07 4824840
add initial cran-comments
