+++ cran-comments.md
@@ -1,0 +1,22 @@
+## Test environments
+
+* local Windows 7, R 3.4.2
+* ubuntu, os x, R 3.4.2 (travis-ci)
+* win-builder (devel and release)
+* Windows 10, R 3.4.2 
+
+## R CMD check results
+
+There were no ERRORs or WARNINGs.
+
+There was 1 NOTE:
+
+Possibly mis-spelled words in DESCRIPTION:
+  Hydrometric (2:34)
+  hydrometric (7:78)
+  
+All words are correctly spelled. 
+
+## Downstream dependencies
+
+There are currently no downstream dependencies.
