### 2017-11-07 efa13e6
@jongoetz vignette edits
