+++ vignettes/tidyhydat_example_analysis.Rmd
@@ -12,12 +12,16 @@ vignette: >
 ---
 
 ```{r setup, include=FALSE}
-knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE, fig.width=7, fig.height=7)
+knitr::opts_chunk$set(echo = TRUE, 
+                      warning = FALSE, 
+                      message = FALSE, 
+                      eval = nzchar(Sys.getenv("hydat_eval")),
+                      fig.width=7, fig.height=7)
 ```
 
 # Introduction
 
-The real power of `tidyhydat` comes with its ability to efficiently import hydrometric data into R which then can be used in conjunction with R's vast array of packages; `tidyhydat` gets hydrometric data into R and allows the user to start conducting data analysis as quickly as possible. This vignette is designed to take a user through two short examples of importing data using `tidyhydat` for use for hydrological visualization and analysis and will therefore use additional tools beyond simply the functions in `tidyhydat`. 
+The real power of `tidyhydat` comes with its ability to efficiently import hydrometric data into R which then can be used in conjunction with R's vast array of packages; `tidyhydat` gets hydrometric data into R and allows the user to start conducting data analysis as quickly as possible. This vignette is designed to take a user through two short examples of importing data using `tidyhydat` for use in hydrological visualization and analysis and will therefore use additional tools beyond simply the functions in `tidyhydat`. 
 
 ## Familiarity with the tidyverse
 `tidyhydat` is designed to exist as a "tidy tool" that is complementary to the [tidyverse](https://www.tidyverse.org/) suite of packages. This vignette assumes some familiarity with basic tidy tools particularly `dplyr`, `tidyr`, `ggplot2` and `lubridate`. The [R for Data Science](http://r4ds.had.co.nz/) book by Garrett Grolemund and Hadley Wickham is an enormously useful book both for general R use and for `tidyhydat`. The following is a list of useful links to begin learning particular aspect of each `tidyverse` package:
@@ -38,18 +42,17 @@ library(lubridate)
 ```
 
 ## Objective
-Use `tidyhydat` to find the longest flow record in the Canadian hydrometric network then use `ggplot2` to create some visualizations of the discharge data. 
+Use `tidyhydat` to find the longest flow record in the Canadian hydrometric network then illustrate the use and ease of `ggplot2` to create some visualizations of the discharge data. 
 
 Your first step is to download the HYDAT database which `tidyhydat` facilitates for you:
 ```{r dl_hy, eval=FALSE}
 download_hydat()
 ```
-This should be the only instance where you will need to interact directly with HYDAT. Each `tidyhydat` function prefixed with `hy` will automatically know where to look for HYDAT saving you the trouble. When using `tidyhydat`, often your first task is to find the station(s) that you are interested in. Because we are interested in the longest record we can extract that information with `hy_stn_data_range()` and then feed that information to `hy_daily_flows()` like this: 
[…]
