### 2017-11-07 35793c8
setting env variable so vignette with run on cran
