### 2017-11-07 #152 @boshek
@sckott Thanks! A few questions:

- I still would like to submit the paper to JOSS. What is the process there? Whatever the case, I would still like to opportunity to review it once more.
- I am interested in doing a blog post. One idea I had was to outline all that I learned. One of the things that I love about R is that no matter what your skill level is, your input can always provide value to someone who is just learning that particular skill. I learned 5 or 6 really useful things when making this package that could be wrapped up into a blog post. Is that something that would be useful? I could also simply submit a "how-to" if that was more desirable.
