### 2017-11-07 #152 @boshek
Also this should probably be changed to "tidyhydat"

https://github.com/orgs/ropensci/teams/tidyhat/members
