### 2017-11-07 #152 @sckott
@boshek Sorry about that, forgot about the JOSS. Will ping back soon with some thoughts.

team name changed.

I like that idea for a blog post!  The community would really benefit from that I think. @stefaniebutland  thoughts on his idea for a post?
