### 2017-11-07 #152 @sckott
What do you mean by publicize a release?  Do you mean to CRAN or on twitter? or somewhere else?
