### 2017-11-07 #152 @stefaniebutland
regarding publicizing, I tweet about new blog post from @rOpenSci when it goes up
