### 2017-11-07 5c0e3ab
update links to ropensci
