+++ paper/paper.bib
@@ -52,7 +52,7 @@
     title = {tidyhydat: Extract and Tidy Canadian Hydrometric Data},
     author = {Sam Albers},
     note = {R package version 0.2.9},
-    url = {https://github.com/bcgov/tidyhydat},
+    url = {https://github.com/ropensci/tidyhydat},
   }
 
 @Comment{jabref-meta: databaseType:bibtex;}
