+++ R/download.R
@@ -264,229 +264,6 @@ realtime_stations <- function(prov_terr_state_loc = NULL) {
   net_tibble
 }
 
-#' Request a token from the Environment and Climate Change Canada [EXPERIMENTAL]
-#' @param username Supplied by ECCC. Defaults to NULL which will use WS_USRNM from .Renviron file
-#' @param password Supplied by ECCC. Defaults to NULL which will use WS_PWD from .Renviron file
-#' Request a token from the ECCC web service using the POST method. This token expires after 10 minutes.
-#' You can only have 5 tokens out at once.
-#'
-#' @details The \code{username} and \code{password} should be treated carefully and should never be entered directly into an r script or console.
-#' Rather these credentials should be stored in your \code{.Renviron} file. The .Renviron file can edited using \code{file.edit("~/.Renviron")}.
-#' In that file, which is only stored locally and is only available to you, you can assign your \code{username} to WS_USRNM and \code{password} 
-#' to WS_PWD and then simply issue \code{token_ws}. See \code{?download_ws} for examples.
-#'
-#' @return The token as a string that should be supplied the \code{download_ws_realtime} function. 
-#' 
-#' 
-#' @family realtime functions
-#' @export
-#'
-
-
-token_ws <- function(username = NULL, password = NULL) {
-  
-  if(is.null(username) | is.null(password)){
-    username <- Sys.getenv("WS_USRNM")
-    password <- Sys.getenv("WS_PWD")
-  }
-  
-  login <- list(
-    username = username,
-    password = password
-  )
-  r <- httr::POST("https://wateroffice.ec.gc.ca/services/auth", body = login)
-  
-  ## A workaround that pauses for the connection 
-  Sys.sleep(2)
-
-  ## If the POST request was not a successful, print out why.
-  ## Possibly could provide errors as per web service guidelines
-  if (httr::status_code(r) == 422) {
-    stop("422 Unprocessable Entity: Username and/or password are missing or are formatted incorrectly.")
-  }
-
-  if (httr::status_code(r) == 403) {
-    stop("403 Forbidden: the web service is denying your request. Try any of the following options: 
-          -Ensure you are not currently using all 5 tokens
-          -Wait a few minutes and try again 
-          -Copy the token_ws call and paste it directly into the console
-          -Try using realtime_ws if you only need water quantity data
-         ")
-  }
-
-  ## Catch all error for anything not covered above.
-  httr::stop_for_status(r)
-
-  message(paste0("This token will expire at ", format(Sys.time() + 10 * 60, "%H:%M:%S")))
-
-  ## Extract token from POST
-  token <- httr::content(r, "text", encoding = "UTF-8")
-
-  token
-}
-
-#' Download realtime data from the ECCC web service [EXPERIMENTAL]
-#' 
-#' Function to actually retrieve data from ECCC web service. Before using this function,
-#' a token from \code{token_ws()} is needed. The maximum number of days that can be 
-#' queried depends on other parameters being requested. If one station is requested, 18 
-#' months of data can be requested. If you continually receiving errors when invoking this
-#' function, reduce the number of observations (via station_number, parameters or dates) being requested. 
-#' 
-#' @param station_number Water Survey of Canada station number.
-#' @param parameters parameter ID. Can take multiple entries. Parameter is a numeric code. See \code{param_id} for options. Defaults to all parameters.
-#' @param start_date Need to be in YYYY-MM-DD. Defaults to 30 days before current date. 
[…]
