+++ data-raw/HYDAT_internal_data/README.md
@@ -1,6 +1,4 @@
-## Water Survey of Canada Sub-Sub-Drainage Areas
-
-- The file `parameter_id.csv` contains Parameter id specification, downloaded from [Environment and Climate Change Canada](http://collaboration.cmc.ec.gc.ca/cmc/hydrometric_additionalData/Document/WebService_Guidelines.pdf) under the ([Open Government License - Canada](http://open.canada.ca/en/open-government-licence-canada)). 
+## tidyhydat internal data
 
 - The file `inst/test_db/tinyhydat.sqlite3` is the first 100 rows (or less) of each table in the hydat database. This is used exclusively for testing purposes and is generated by `tinyhydat_proc.R`
 

[…]
