### 2017-11-07 57c3eca
remove built vignettes
