### 2017-11-08 #152 @sckott
@boshek for JOSS thing, can you

* generate a new release with a DOI. this doesn't mean you have to submit to CRAN of course
* Submit to JOSS at http://joss.theoj.org/papers/new
