### 2017-11-08 3e08625
revamp travis.yml
