### 2017-11-08 d803786
removing built vignettes
