### 2017-11-17 b55ac0c
add reviewers to DESCRIPTION
