### 2017-12-06 9954d53
update JOSS paper
