### 2017-12-14 ba040c5
update README with doi
