### 2017-12-14 109ab5d
final version of paper.md
