+++ inst/JOSS/paper.bib
@@ -31,6 +31,13 @@
   publisher={Public Library of Science}
 }
 
+@misc{wickham2016r,
+  title={R for data science},
+  author={Wickham, Hadley and Grolemund, Garrett},
+  year={2016},
+  publisher={Sebastopol, CA: O’Reilly. http:// r4ds. had. co. nz}
+}
+
 @Manual{RCore,
   title = {R: A Language and Environment for Statistical Computing},
   author = {{R Core Team}},
@@ -51,7 +58,8 @@
   @Manual{alberstidyhydat,
     title = {tidyhydat: Extract and Tidy Canadian Hydrometric Data},
     author = {Sam Albers},
-    note = {R package version 0.2.9},
+    year = {2017},
+    note = {R package version 0.3.0},
     url = {https://github.com/ropensci/tidyhydat},
   }
 
