### 2017-12-14 25f9dd3
Add a means to check if a new hydat should be downloaded. Closes #61
