+++ R/zzz.R
@@ -109,18 +109,27 @@
   if(!file.exists(file.path(hy_dir(),"Hydat.sqlite3"))){
     packageStartupMessage("Tidyhydat requires HYDAT which has not yet been downloaded. Download using download_hydat()")
   }
-#  
-#  ## Check the date of current HYDAT
-#  base_url <- "http://collaboration.cmc.ec.gc.ca/cmc/hydrometrics/www/"
-#  x <- httr::GET(base_url)
-#  httr::stop_for_status(x)
-#  new_hydat <- substr(gsub(
-#    "^.*\\Hydat_sqlite3_", "",
-#    httr::content(x, "text")
-#  ), 1, 8)
-#
-#  # Do we need to download a new version?
-#  if (!(as.Date(new_hydat, "%Y%m%d") == as.Date(hy_version()$Date))) {
-#    packageStartupMessage(paste0("A new version of HYDAT was release on ", as.Date(new_hydat, "%Y%m%d"),". Use download_hydat() to get the new version."))
-#  }
+  
+  ## HYDAT is updated quarterly - should we go check if a new one is available for download?
+  ## Only check when there is likely a new version i.e. about 3 months after last version
+  if(Sys.Date() > (as.Date(hy_version()$Date) + 10)){
+    packageStartupMessage("Checking for a new version of HYDAT...")
+    base_url <- "http://collaboration.cmc.ec.gc.ca/cmc/hydrometrics/www/"
+    x <- httr::GET(base_url)
+    httr::stop_for_status(x)
+    
+    ## Extract newest HYDAT
+    new_hydat <- as.Date(substr(gsub(
+      "^.*\\Hydat_sqlite3_", "",
+      httr::content(x, "text")
+    ), 1, 8),  "%Y%m%d")
+    
+    ## Compare that to existing HYDAT
+    if (new_hydat != as.Date(hy_version()$Date)){
+      packageStartupMessage(paste0("Your version of HYDAT is out of date. Use download_hydat() to get the new version."))
+    } else{
+      packageStartupMessage("No new version of HYDAT")
+    }
+  }
+
 }
