### 2017-12-14 c062897
add single quotes to jargon in DESCRIPTION
