### 2017-12-14 b9ed9aa
.onAttach checks if HYDAT even exists before trying to use it
