### 2017-12-15 2410147
use actual SQL for hy_stations. Starting to address #67
