### 2017-12-15 3bdc5c4
Update version, news, description for CRAN release
