+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: tidyhydat
 Title: Extract and Tidy Canadian 'Hydrometric' Data
-Version: 0.3.0
+Version: 0.3.1
 Authors@R: c(person("Sam", "Albers", email = "sam.albers@gov.bc.ca", role = c("aut", "cre")),
     person("David", "Hutchinson", email = "david.hutchinson@canada.ca", role = "ctb"),
     person("Province of British Columbia", role = "cph"),
@@ -19,8 +19,8 @@ BugReports: https://github.com/ropensci/tidyhydat/issues
 Encoding: UTF-8
 Imports:
     dplyr (>= 0.7.3),
-    dbplyr (>= 1.1.0),
     readr (>= 1.1.1),
+    dbplyr (>= 1.1.0),
     lubridate (>= 1.6.0),
     tidyr (>= 0.7.1),
     DBI (>= 0.7),
