### 2017-12-15 #152 @noamross
@boshek Question: I see that you put this on CRAN with the `rev` category for your contributors.  Did you get a CRAN notes when doing so, or did you build and submit using `r-devel`? We've been quietly encouraging this but not too much because there were some snafus with CRAN submission.  If it went through with no issues that's a nice milestone.
