### 2017-12-15 #152 @boshek
@noamross No snafus! Some UTF-8 string notes that I haven't figured out how to deal with yet but the `rev` category worked fine.  I included the link to the onboarding issue but I think that I would omit that as it quite ugly on the [CRAN package page](https://CRAN.R-project.org/package=tidyhydat).
