### 2017-12-15 #152 @boshek
Removing the links is particularly important because I actually linked to wrong issue. GAH!
