### 2017-12-20 #152 @karthik
Looking over your README, it does show as peer reviewed to me as well.

![image](https://user-images.githubusercontent.com/138494/34219787-2c678ef2-e567-11e7-9b70-4b7254835d47.png)
